Archaeological Data

A virtual S4 class to represent archaeological data.

Slots

samples

A character vector.

groups

A character vector.

totals

A numeric vector giving the absolute row sums.

tpq

An integer vector specifying the TPQ of each row.

taq

An integer vector specifying the TAQ of each row.

Get and set

In the code snippets below, x is an AbundanceMatrix object.

get_samples(x) and get_samples(x) <- value

Get or set the sample names of x.

get_groups(x) and set_groups(x) <- value

Get or set the groups of x.

get_dates(x) and set_dates(x) <- value

Get or set the dates of x. value can be a list with components tpq (TPQ - terminus post quem) and taq (TAQ - terminus ante quem) or a numeric vector (point estimate). The elements of value are coerced to integer with as.integer() (hence truncated towards zero).

Chronology

The way chronological information is handled is somewhat opinionated:

Sub-annual precision is overkill/meaningless in most situations: dates are assumed to be expressed in years CE and are stored as integer (values are coerced with as.integer() and hence truncated towards zero).

A date range (i.e. TPQ/TAQ) is expected, although it is possible to specify a point estimate.

Author

N. Frerebeau