Correspondence Analysis Results

An S4 class to store the results of a simple correspondence analysis.

Slots

dimension

An integer giving the dimension of the solution.

row_names

A character vector specifying the row names.

row_coordinates

A numeric matrix giving the row standard coordinates.

row_distances

A numeric vector giving the row chi-square distances to centroid.

row_inertias

A numeric vector giving the row inertias.

row_masses

A numeric vector giving the row masses.

row_supplement

A logical vector specifying if a row is a supplementary point.

column_names

A character vector specifying the column names.

column_coordinates

A numeric matrix giving the column standard coordinates.

column_distances

A numeric vector giving the column chi-square distances to centroid.

column_inertias

A numeric vector giving the column inertias.

column_masses

A numeric vector giving the column masses.

column_supplement

A logical vector specifying if a column is a supplementary point.

singular_values

A numeric vector giving the singular values.

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

#>                CA1         CA2          CA3         CA4          CA5  .sup
#> row1   0.149485834 -0.50821842 -0.135330154 -0.02484872 -0.027496041 FALSE
#> row2   0.678447792 -0.07840110  0.089371456 -0.01814593  0.017970498 FALSE
#> row3  -0.460475426 -0.01813191 -0.132786185 -0.14756008  0.017725885 FALSE
#> row4  -0.258606350 -0.08879884 -0.009988114  0.21316960  0.024527735 FALSE
#> row5  -0.008872788  0.34496233 -0.006033888  0.03393694 -0.057838400 FALSE
#> row6  -0.337587329 -0.12201507  0.393469653 -0.08047755 -0.001496124 FALSE
#> row7   0.186186562  0.41119365 -0.079195403 -0.04222683  0.037406201 FALSE
#> row8   0.580336566  0.31186919  0.171438612 -0.17577993 -0.004657920  TRUE
#> row9  -0.474394024 -0.23154403 -0.039906526  0.17262859 -0.112877204  TRUE
#> row10  0.199783841  0.07534443  0.394489622 -0.18226291  0.071702871  TRUE

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

#>              CA1         CA2         CA3          CA4          CA5  .sup
#> col1   0.1782952  0.12966036 -0.08655951 -0.212871858  0.006069145 FALSE
#> col2   0.4492730  0.41148007  0.07769235  0.114474061  0.016014426 FALSE
#> col3  -0.3601879  0.18926899 -0.10838133  0.050691800 -0.051131032 FALSE
#> col4   0.3946244 -0.47778163  0.11532216  0.009253878 -0.035667975 FALSE
#> col5  -0.5175174 -0.01226027  0.25516856 -0.033475665  0.022165002 FALSE
#> col6  -0.1057465 -0.26906480 -0.18311956  0.075882537  0.039489790 FALSE
#> col7  -0.3206892 -0.35189851 -0.36203956 -0.005438329  0.090474620  TRUE
#> col8  -0.3903889 -0.12784114 -0.13710739 -0.101138543  0.287468374  TRUE
#> col9  -0.1434572  0.01290283 -0.03638382  0.044975325  0.036476598  TRUE
#> col10 -0.1705237 -0.24953423  0.18527499 -0.215817109  0.021406681  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>      row1      row2      row3      row4      row5      row6      row7      row8 
#> 0.5480143 0.6892587 0.5020824 0.3477141 0.3515838 0.5386558 0.4617357 0.7031105 
#>      row9     row10 
#> 0.5681521 0.4894625 

## Get row inertias
get_inertia(X, margin = 1)

#>       row1       row2       row3       row4       row5       row6       row7 
#> 0.04325593 0.06842682 0.03630879 0.01990208 0.02136489 0.02865680 0.02807571 
#>       row8       row9      row10 
#>         NA         NA         NA 

## Get row contributions
get_contributions(X, margin = 1)

#>              CA1        CA2         CA3        CA4         CA5
#> row1  2.53631565 44.4272527 11.74089253  0.7520171 10.25004281
#> row2 52.24396057  1.0572872  5.12046767  0.4010312  4.37830047
#> row3 24.06667772  0.0565504 11.30362516 26.5190114  4.25991740
#> row4  8.67508105  1.5500838  0.07309218 63.2502202  9.32163640
#> row5  0.01072271 24.5625391  0.02800836  1.6832405 54.42508513
#> row6  8.86989950  1.7559768 68.05775885  5.4089405  0.02080961
#> row7  3.59734279 26.5903101  3.67615525  1.9855391 17.34420818

## Get eigenvalues
get_eigenvalues(X)

#>      eigenvalues   percentage cumulative
#> CA1 1.268991e-01 5.158689e+01   51.58689
#> CA2 8.373617e-02 3.404034e+01   85.62722
#> CA3 2.246724e-02 9.133360e+00   94.76058
#> CA4 1.182612e-02 4.807541e+00   99.56813
#> CA5 1.062372e-03 4.318741e-01  100.00000
#> CA6 1.891555e-34 7.689528e-32  100.00000