Data Matrix

A virtual S4 class to represent a data matrix.

Arguments

data
A data vector.

nrow
An integer value giving the desired number of rows.

ncol
An integer value giving the desired number of columns.

byrow
A logical scalar: should the matrix be filled by rows? If FALSE (the default) the matrix is filled by columns.

dimnames
A list of length 2 giving the row and column names respectively. If NULL (the default) dimension names will be created.

Slots

values

An integer, a numeric or a logical vector (see details).

Access

In the code snippets below, x is a *Matrix object.

length(x)

Returns the length of x.

dim(x)

Returns the dimension of x.

nrow(x)

Returns the number of rows present in x.

ncol(x)

Returns the number of columns present in x.

dimnames(x), dimnames(x) <- value

Retrieves or sets the row dimnames of x according to value.

rownames(x), rownames(x) <- value

Retrieves or sets the row names of x according to value.

colnames(x), colnames(x) <- value

Retrieves or sets the column names of x according to value.

Subset

In the code snippets below, x is a *Matrix object.

x[i, j, ..., drop]

Extracts elements selected by subscripts i and j. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the dimnames of the object. An empty index (a comma separated blank) indicates that all entries in that dimension are selected. Returns an object of the same class as x.

x[[i]]

Extracts a single element selected by subscript i.

See also

Author

N. Frerebeau