arkhe: Representation of Archaeological Data

A collection of classes that represent archaeological data. This package provides a set of S4 classes that represent different special types of matrix (absolute/relative frequency, presence/absence data, co-occurrence matrix, etc.) upon which package developers can build subclasses. It also provides a set of generic methods (mutators and coercion mechanisms) and functions (e.g. summary statistics, predicates). In addition, a few classes of general interest (e.g. that represent stratigraphic relationships) are implemented.

Details

Package:	arkhe

Type:	Package

Version:	0.3.1

License:	GPL-3

doi: 10.5281/zenodo.3526659

Package options

arkhe uses the following options() to configure behavior:

arkhe.autodetect: a logical scalar.

arkhe.verbose: a logical scalar. Should R report extra information on progress?

See also

Author

Full list of authors and contributors (alphabetic order):

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

IRAMAT-CRP2A (UMR 5060)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France