Chronological Information

set_dates and get_dates allow to get or set the dates of an object.

get_dates(object)

set_dates(object) <- value

# S4 method for GenericMatrix
get_dates(object)

# S4 method for GenericMatrix
set_dates(object) <- value

Arguments

object
A \(m \times p\) matrix of count data (typically of class Matrix).

value
A possible value for the element(s) of object (see below).

Value

get_dates returns a two-columns data.frame.

Details

An attempt is made to interpret the argument value in a suitable way. Note that errors are assumed to be given at 1 sigma.

If value is a:

character vector

it is assumed to contain Roman numerals.

numeric or integer vector

these values are assumed to represent dates without known errors.

list

containing components "value" and "error", these are used to define dates and corresponding errors.

matrix or data.frame with two or more columns

the first is assumed to contain the dates and the second the error values. Note that if value has columns named "value" and "error", these columns will be used.

Author

N. Frerebeau

Examples

## Create a count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 10, ncol = 10,
                 dimnames = list(LETTERS[1:10], NULL))

## Set dates as a matrix
B <- matrix(data = sample(0:10, 20, TRUE), nrow = 10, ncol = 2)
set_dates(A) <- B
get_dates(A)

#>   value error
#> A     6     7
#> B     2     2
#> C     7     6
#> D     2     6
#> E     5     7
#> F     5     7
#> G     7     3
#> H     2     8
#> I     4     5
#> J     3     1

## Set dates as a data.frame
B <- data.frame(sample(0:10, 5, TRUE), sample(0:10, 5, TRUE),
                row.names = LETTERS[seq(1, 10, 2)])
set_dates(A) <- B
get_dates(A)

#>   value error
#> A     5     0
#> B    NA    NA
#> C     9     1
#> D    NA    NA
#> E    10     0
#> F    NA    NA
#> G     7     2
#> H    NA    NA
#> I     4     5
#> J    NA    NA

## Set dates as a list
B <- list(value = sample(0:10, 10, TRUE), error = sample(0:10, 10, TRUE))
set_dates(A) <- B
get_dates(A)

#>   value error
#> A     8     4
#> B     1     7
#> C     5     9
#> D     7     3
#> E     4     6
#> F     2    10
#> G     9     3
#> H    10     9
#> I     8     5
#> J     6     0

## Set dates as a numeric vector
B <- sample(0:10, 10, TRUE)
suppressWarnings(set_dates(A) <- B)
get_dates(A)

#>   value error
#> A     5    NA
#> B     5    NA
#> C    10    NA
#> D     1    NA
#> E     5    NA
#> F     7    NA
#> G    10    NA
#> H     5    NA
#> I     2    NA
#> J     6    NA

## Set dates as a character vector
B <- c("X", "IX", "VIII", "VII", "VI", "V", "IV", "III", "II", "I")
set_dates(A) <- B
get_dates(A)

#>   value error
#> A    10    NA
#> B     9    NA
#> C     8    NA
#> D     7    NA
#> E     6    NA
#> F     5    NA
#> G     4    NA
#> H     3    NA
#> I     2    NA
#> J     1    NA

## Unset dates
set_dates(A) <- NULL
get_dates(A)

#> [1] value error
#> <0 rows> (or 0-length row.names)