Correlation, Variance and Covariance

Computes the variance, covariance or correlation of a DataMatrix.

# S4 method for DataMatrix,missing
var(x, na.rm = FALSE)

# S4 method for DataMatrix,missing
cov(x, use = "everything", method = c("pearson", "kendall", "spearman"))

# S4 method for DataMatrix,missing
cor(x, use = "everything", method = c("pearson", "kendall", "spearman"))

Arguments

x
A DataMatrix object.

na.rm
A logical scalar: should missing values (including NaN) be omitted from the calculations?

use
An optional character string giving a method for computing covariances in the presence of missing values (see cor).

method
A character string indicating which correlation coefficient (or covariance) is to be computed (see cor).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column means
rowMeans(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>   5.6   5.4   7.2   3.4   3.6   5.0   5.8   3.2   6.0   5.2   6.0   4.4   6.8 
#> row14 row15 row16 row17 row18 row19 row20 
#>   5.6   4.8   6.0   5.8   6.8   5.6   4.6 

colMeans(A)

#> col1 col2 col3 col4 col5 
#> 4.75 5.15 5.60 5.80 5.40 

## Row and column sums
rowSums(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>    28    27    36    17    18    25    29    16    30    26    30    22    34 
#> row14 row15 row16 row17 row18 row19 row20 
#>    28    24    30    29    34    28    23 

colSums(A)

#> col1 col2 col3 col4 col5 
#>   95  103  112  116  108 

## Variance, covariance and correlation matrix
var(A)

#>            col1       col2       col3      col4      col5
#> col1  4.8289474  0.9342105 -0.2631579  1.105263  1.052632
#> col2  0.9342105 11.0815789 -0.9368421 -1.126316 -2.063158
#> col3 -0.2631579 -0.9368421  8.8842105 -4.715789 -1.042105
#> col4  1.1052632 -1.1263158 -4.7157895  7.431579  1.978947
#> col5  1.0526316 -2.0631579 -1.0421053  1.978947  8.042105

cov(A)

#>            col1       col2       col3      col4      col5
#> col1  4.8289474  0.9342105 -0.2631579  1.105263  1.052632
#> col2  0.9342105 11.0815789 -0.9368421 -1.126316 -2.063158
#> col3 -0.2631579 -0.9368421  8.8842105 -4.715789 -1.042105
#> col4  1.1052632 -1.1263158 -4.7157895  7.431579  1.978947
#> col5  1.0526316 -2.0631579 -1.0421053  1.978947  8.042105

cor(A)

#>            col1       col2       col3       col4       col5
#> col1  1.0000000  0.1277079 -0.0401773  0.1845011  0.1689139
#> col2  0.1277079  1.0000000 -0.0944183 -0.1241134 -0.2185480
#> col3 -0.0401773 -0.0944183  1.0000000 -0.5803693 -0.1232870
#> col4  0.1845011 -0.1241134 -0.5803693  1.0000000  0.2559816
#> col5  0.1689139 -0.2185480 -0.1232870  0.2559816  1.0000000