Correspondence Analysis

Computes a simple correspondence analysis based on the singular value decomposition.

ca(object, ...)

get_coordinates(object, ...)

get_contributions(object, ...)

get_distances(object, ...)

get_eigenvalues(object)

get_inertia(object, ...)

# S4 method for CountMatrix
ca(object, n = min(dim(object) - 1), sup_rows = NULL, sup_columns = NULL)

# S4 method for CA
predict(object, data, margin = 1)

# S4 method for CA
get_coordinates(
  object,
  margin = 1,
  standard = FALSE,
  sup = TRUE,
  sup_name = ".sup"
)

# S4 method for CA
get_inertia(object, margin = 1)

# S4 method for CA
get_distances(object, margin = 1)

# S4 method for CA
get_contributions(object, margin = 1)

# S4 method for CA
get_eigenvalues(object)

Arguments

object
A CountMatrix or a CA object.

...
Currently not used.

n
An integer value specifying the number of dimensions to be kept in the results (default to 5).

sup_rows
A numeric or logical vector specifying the indices of the supplementary rows.

sup_columns
A numeric or logical vector specifying the indices of the supplementary columns.

data
An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates rows (the default), 2 indicates columns.

standard
A logical scalar: should standard coordinates be returned instead of principal coordinates (the default)?

sup
A logical scalar: should supplementary points be returned?

sup_name
A character string specifying the name of the column to create for supplementary points attribution (see below).

Value

ca returns a CA object.

predict returns a data.frame of principal coordinates.

get_coordinates returns a data.frame of coordinates. If sup is TRUE, an extra column (named after sup_name) is added specifying whether an observation is a supplementary point or not.

get_inertia returns a data.frame.

get_distances returns a data.frame.

get_contributions returns a data.frame.

get_eigenvalues returns a data.frame with the following columns: eigenvalues, percentage (percentage of variance) and cumulative (cumulative percentage of variance).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

#>               CA1         CA2         CA3           CA4          CA5  .sup
#> row1  -0.59793400  0.02311985 -0.22173614 -0.0225929706  0.012365165 FALSE
#> row2   0.06924897 -0.16821107  0.07923714 -0.1903558923 -0.041187663 FALSE
#> row3   0.11042978  0.41374923  0.16989414 -0.0759062505  0.060804969 FALSE
#> row4   0.18739403 -0.09254562 -0.10479707  0.0006301948  0.075693218 FALSE
#> row5  -0.27802363 -0.20677666  0.26713779  0.1493200023  0.019338981 FALSE
#> row6   0.02957397  0.19580564 -0.01545791  0.0934528748 -0.083156848 FALSE
#> row7   0.56564504 -0.09786430 -0.14635751  0.0839775475  0.003205579 FALSE
#> row8  -0.31550914 -0.05611258  0.12160459  0.2042344603 -0.243467759  TRUE
#> row9   0.15915746 -0.05682938  0.08768751 -0.2656492535 -0.240478031  TRUE
#> row10 -0.19077508  0.80593061 -0.29155015  0.1883550603 -0.258372237  TRUE

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

#>                CA1          CA2         CA3         CA4           CA5  .sup
#> col1  -0.417652234  0.077013550  0.01140891 -0.07172461  0.0980896087 FALSE
#> col2   0.499704067 -0.027940016 -0.12570178 -0.16817672 -0.0003863714 FALSE
#> col3  -0.118844220  0.417059426 -0.02381107  0.02275026 -0.0573530879 FALSE
#> col4  -0.486254598 -0.271447890  0.02946775 -0.08021790 -0.0732371143 FALSE
#> col5   0.257015603 -0.040995272  0.28589256  0.05628556  0.0058385708 FALSE
#> col6   0.050873490 -0.108519837 -0.17721375  0.16262727  0.0109805976 FALSE
#> col7   0.044675354 -0.271869353 -0.02884261  0.05699291 -0.0581376643  TRUE
#> col8  -0.004794749  0.124514403  0.22050793 -0.04099506  0.1472645356  TRUE
#> col9  -0.205831532  0.117616969  0.08945914  0.17154530  0.0887241387  TRUE
#> col10  0.374717297 -0.004981155 -0.21657167  0.21086173  0.0418882853  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>      row1      row2      row3      row4      row5      row6      row7      row8 
#> 0.6386625 0.2780300 0.4708569 0.2457509 0.4626951 0.2347380 0.5983433 0.4674084 
#>      row9     row10 
#> 0.4057696 0.9344274 

## Get row inertias
get_inertia(X, margin = 1)

#>        row1        row2        row3        row4        row5        row6 
#> 0.062248250 0.013970006 0.024930816 0.007518869 0.028372947 0.009073006 
#>        row7        row8        row9       row10 
#> 0.047447739          NA          NA          NA 

## Get row contributions
get_contributions(X, margin = 1)

#>             CA1        CA2        CA3          CA4         CA5
#> row1 47.8766877  0.2104586 29.3249840 6.181446e-01  0.88100679
#> row2  0.7604544 13.1927340  4.4345685 5.196429e+01 11.57557705
#> row3  1.2032739 49.6644452 12.6851634 5.141300e+00 15.69755477
#> row4  3.8362525  2.7509747  5.3436893 3.923477e-04 26.93217739
#> row5  8.9889995 14.6194199 36.9628806 2.344821e+01  1.87144446
#> row6  0.1263680 16.2872365  0.1537681 1.141114e+01 42.99082066
#> row7 37.2079640  3.2747312 11.0949460 7.416520e+00  0.05141889

## Get eigenvalues
get_eigenvalues(X)

#>     eigenvalues percentage cumulative
#> CA1  0.11396373  58.877230   58.87723
#> CA2  0.03876033  20.024802   78.90203
#> CA3  0.02558700  13.219047   92.12108
#> CA4  0.01260204   6.510608   98.63169
#> CA5  0.00264853   1.368313  100.00000