Count values using a predicate

Counts values by rows/columns using a predicate function.

Usage

count(x, f, ...)

# S4 method for matrix,`function`
count(x, f, margin = 1, negate = FALSE)

# S4 method for data.frame,`function`
count(x, f, margin = 1, negate = FALSE)

Arguments

x

An object (should be a matrix or a data.frame).

f

A predicate function.

...

Currently not used.

margin

A vector giving the subscripts which the function will be applied over (1 indicates rows, 2 indicates columns).

negate

A logical scalar: should the negation of f be used instead of f?

Value

A numeric vector.

See also

Author

N. Frerebeau

Examples

## Create a count data matrix
X <- matrix(sample(1:10, 25, TRUE), nrow = 5, ncol = 5)

## Add NA
k <- sample(1:25, 3, FALSE)
X[k] <- NA
X
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    6    8   NA    8    8
#> [2,]    9   10    9    2    6
#> [3,]    3    7    6    3    9
#> [4,]    4    3   NA    1    7
#> [5,]   10    4   NA    5    4

## Count missing values in rows
count(X, f = is.na, margin = 1)
#> [1] 1 0 0 1 1
## Count non-missing values in columns
count(X, f = is.na, margin = 2, negate = TRUE)
#> [1] 5 5 2 5 5

## Find row with NA
detect(X, f = is.na, margin = 1)
#> [1]  TRUE FALSE FALSE  TRUE  TRUE
## Find column without any NA
detect(X, f = is.na, margin = 2, negate = TRUE, all = TRUE)
#> [1]  TRUE  TRUE FALSE  TRUE  TRUE

## Keep row without any NA
keep(X, f = is.na, margin = 1, negate = TRUE, all = TRUE)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    9   10    9    2    6
#> [2,]    3    7    6    3    9
## Keep row without any NA
keep(X, f = is.na, margin = 2, negate = TRUE, all = TRUE)
#>      [,1] [,2] [,3] [,4]
#> [1,]    6    8    8    8
#> [2,]    9   10    2    6
#> [3,]    3    7    3    9
#> [4,]    4    3    1    7
#> [5,]   10    4    5    4

## Remove row with any NA
discard(X, f = is.na, margin = 1, all = FALSE)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    9   10    9    2    6
#> [2,]    3    7    6    3    9
## Remove column with any NA
discard(X, f = is.na, margin = 2, all = FALSE)
#>      [,1] [,2] [,3] [,4]
#> [1,]    6    8    8    8
#> [2,]    9   10    2    6
#> [3,]    3    7    3    9
#> [4,]    4    3    1    7
#> [5,]   10    4    5    4

## Replace NA with zeros
replace_NA(X, value = 0)
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    6    8    0    8    8
#> [2,]    9   10    9    2    6
#> [3,]    3    7    6    3    9
#> [4,]    4    3    0    1    7
#> [5,]   10    4    0    5    4