Get or Set Parts of an Object

Getters and setters to retrieve or set parts of an object.

has_groups(x)

get_groups(x)

set_groups(x) <- value

get_samples(x)

set_samples(x) <- value

has_dates(x)

get_dates(x)

set_dates(x) <- value

get_tpq(x)

set_tpq(x) <- value

get_taq(x)

set_taq(x) <- value

get_totals(x)

set_totals(x) <- value

# S4 method for AbundanceMatrix
has_groups(x)

# S4 method for AbundanceMatrix
get_groups(x)

# S4 method for AbundanceMatrix
get_samples(x)

# S4 method for AbundanceMatrix
has_dates(x)

# S4 method for AbundanceMatrix
get_dates(x)

# S4 method for AbundanceMatrix
get_tpq(x)

# S4 method for AbundanceMatrix
get_taq(x)

# S4 method for AbundanceMatrix
get_totals(x)

# S4 method for OccurrenceMatrix
get_totals(x)

# S4 method for AbundanceMatrix
set_groups(x) <- value

# S4 method for AbundanceMatrix
set_samples(x) <- value

# S4 method for AbundanceMatrix,`NULL`
set_dates(x) <- value

# S4 method for AbundanceMatrix,numeric
set_dates(x) <- value

# S4 method for AbundanceMatrix,list
set_dates(x) <- value

# S4 method for AbundanceMatrix,`NULL`
set_tpq(x) <- value

# S4 method for AbundanceMatrix,numeric
set_tpq(x) <- value

# S4 method for AbundanceMatrix,`NULL`
set_taq(x) <- value

# S4 method for AbundanceMatrix,numeric
set_taq(x) <- value

# S4 method for AbundanceMatrix
set_totals(x) <- value

Arguments

x

An object from which to get or set element(s) (typically an AbundanceMatrix object).

value

A possible value for the element(s) of x.

Value

set_*() returns an object of the same sort as x with the new values assigned.

get_*() returns the part of x.

has_*() returns a logical scalar.

Details

get_samples(x) and get_samples(x) <- value

Get or set the sample names of x.

get_groups(x) and set_groups(x) <- value

Get or set the groups of x.

get_dates(x) and set_dates(x) <- value

Get or set the dates of x. value can be a list with components tpq (TPQ - terminus post quem) and taq (TAQ - terminus ante quem) or a numeric vector (point estimate). The elements of value are coerced to integer with as.integer() (hence truncated towards zero).

See also

Author

N. Frerebeau

Examples

## Create a data.frame
X <- matrix(data = sample(0:10, 50, TRUE), nrow = 10, ncol = 5)
Y <- as.data.frame(X)

## Coerce to a count matrix
Z <- as_count(Y)

## Set/get groups
set_samples(Z) <- rep(c("a", "b", "c", "d", "e"), each = 2)
get_samples(Z)
#>  [1] "a" "a" "b" "b" "c" "c" "d" "d" "e" "e"

## Set/get groups
set_groups(Z) <- rep(c("A", "B"), each = 5)
get_groups(Z)
#>  [1] "A" "A" "A" "A" "A" "B" "B" "B" "B" "B"

## Get/get dates
dates <- list(
  tpq = sample(1301:1400, 10, replace = TRUE),
  taq = sample(1451:1500, 10, replace = TRUE)
)
set_dates(Z) <- dates
get_dates(Z)
#>     tpq  taq
#> 1  1338 1473
#> 2  1370 1490
#> 3  1311 1473
#> 4  1312 1471
#> 5  1348 1488
#> 6  1306 1486
#> 7  1388 1470
#> 8  1391 1491
#> 9  1348 1486
#> 10 1399 1461

## Collection of features
as_features(Z)
#>    V1 V2 V3 V4 V5 samples group  tpq  taq
#> 1   6  7  5  8  4       a     A 1338 1473
#> 2   2  2  9  1  7       a     A 1370 1490
#> 3   7  6 10  5  9       b     A 1311 1473
#> 4   2  6  7  7  3       b     A 1312 1471
#> 5   5  7  4  4  6       c     A 1348 1488
#> 6   5  7  0  2 10       c     B 1306 1486
#> 7   7  3  1  9  3       d     B 1388 1470
#> 8   2  8  0 10  9       d     B 1391 1491
#> 9   4  5  2  8  5       e     B 1348 1486
#> 10  3  1  5  6  0       e     B 1399 1461