Common Operations on Matrix Objects

Performs common operations on DataMatrix objects.

# S4 method for DataMatrix,DataMatrix
Arith(e1, e2)

# S4 method for DataMatrix,numeric
Arith(e1, e2)

# S4 method for DataMatrix,DataMatrix
Compare(e1, e2)

# S4 method for DataMatrix,numeric
Compare(e1, e2)

# S4 method for DataMatrix,DataMatrix
Logic(e1, e2)

# S4 method for DataMatrix,numeric
Logic(e1, e2)

# S4 method for DataMatrix,logical
Logic(e1, e2)

# S4 method for DataMatrix
Summary(x, na.rm = FALSE)

Arguments

x, e1, e2
An object (typically a DataMatrix object).

na.rm
A logical scalar: should missing values (including NaN) be omitted from the calculations?

Group Generics

DataMatrix objects have support for S4 group generic functionality to operate within elements across objects:

Arith

"+", "-", "*", "^", "%%", "%/%", "/"

Compare

"==", ">", "<", "!=", "<=", ">="

Logic

"&", "|"

Ops

"Arith", "Compare", "Logic"

Summary

"min", "max", "range", "prod", "sum", "any", "all"

Other Methods

rowAll and colAll

Check if all values are TRUE according to a given predicate.

rowAny and colAny

Check if some values are TRUE according to a given predicate.

rowCount and colCount

Count values in rows or columns according to a given predicate.

Examples

## Create count data matrix
A <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
B <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Compare
A == B

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE FALSE FALSE FALSE FALSE
#>  [2,] FALSE FALSE FALSE FALSE FALSE
#>  [3,] FALSE FALSE FALSE  TRUE FALSE
#>  [4,] FALSE FALSE FALSE  TRUE  TRUE
#>  [5,] FALSE FALSE FALSE FALSE FALSE
#>  [6,] FALSE FALSE FALSE FALSE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE
#>  [9,] FALSE FALSE FALSE FALSE FALSE
#> [10,] FALSE FALSE FALSE  TRUE FALSE
#> [11,] FALSE FALSE FALSE FALSE  TRUE
#> [12,] FALSE FALSE FALSE FALSE FALSE
#> [13,] FALSE FALSE FALSE FALSE  TRUE
#> [14,] FALSE FALSE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE FALSE FALSE
#> [16,] FALSE FALSE FALSE FALSE FALSE
#> [17,] FALSE FALSE FALSE FALSE FALSE
#> [18,] FALSE FALSE  TRUE FALSE FALSE
#> [19,] FALSE FALSE FALSE FALSE FALSE
#> [20,]  TRUE FALSE FALSE FALSE FALSE

A != B

#>        [,1] [,2]  [,3]  [,4]  [,5]
#>  [1,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [2,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [3,]  TRUE TRUE  TRUE FALSE  TRUE
#>  [4,]  TRUE TRUE  TRUE FALSE FALSE
#>  [5,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [6,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [7,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [8,]  TRUE TRUE  TRUE  TRUE  TRUE
#>  [9,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [10,]  TRUE TRUE  TRUE FALSE  TRUE
#> [11,]  TRUE TRUE  TRUE  TRUE FALSE
#> [12,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [13,]  TRUE TRUE  TRUE  TRUE FALSE
#> [14,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [15,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [16,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [17,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [18,]  TRUE TRUE FALSE  TRUE  TRUE
#> [19,]  TRUE TRUE  TRUE  TRUE  TRUE
#> [20,] FALSE TRUE  TRUE  TRUE  TRUE

A > 5

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,]  TRUE FALSE  TRUE FALSE  TRUE
#>  [2,]  TRUE  TRUE FALSE FALSE FALSE
#>  [3,] FALSE  TRUE  TRUE  TRUE  TRUE
#>  [4,] FALSE FALSE FALSE FALSE FALSE
#>  [5,]  TRUE  TRUE FALSE FALSE  TRUE
#>  [6,] FALSE  TRUE FALSE FALSE  TRUE
#>  [7,] FALSE  TRUE FALSE  TRUE FALSE
#>  [8,]  TRUE  TRUE FALSE FALSE  TRUE
#>  [9,] FALSE  TRUE FALSE  TRUE FALSE
#> [10,] FALSE  TRUE FALSE FALSE FALSE
#> [11,] FALSE  TRUE FALSE  TRUE  TRUE
#> [12,]  TRUE FALSE  TRUE  TRUE FALSE
#> [13,]  TRUE FALSE  TRUE  TRUE  TRUE
#> [14,] FALSE  TRUE  TRUE FALSE  TRUE
#> [15,]  TRUE  TRUE  TRUE FALSE FALSE
#> [16,]  TRUE FALSE FALSE FALSE  TRUE
#> [17,]  TRUE FALSE FALSE FALSE  TRUE
#> [18,]  TRUE  TRUE FALSE  TRUE FALSE
#> [19,] FALSE FALSE  TRUE  TRUE  TRUE
#> [20,]  TRUE  TRUE  TRUE FALSE FALSE

A < 5

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE  TRUE FALSE  TRUE FALSE
#>  [2,] FALSE FALSE  TRUE  TRUE  TRUE
#>  [3,]  TRUE FALSE FALSE FALSE FALSE
#>  [4,]  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [5,] FALSE FALSE  TRUE  TRUE FALSE
#>  [6,]  TRUE FALSE FALSE  TRUE FALSE
#>  [7,]  TRUE FALSE  TRUE FALSE  TRUE
#>  [8,] FALSE FALSE  TRUE  TRUE FALSE
#>  [9,]  TRUE FALSE  TRUE FALSE  TRUE
#> [10,]  TRUE FALSE  TRUE  TRUE FALSE
#> [11,]  TRUE FALSE  TRUE FALSE FALSE
#> [12,] FALSE  TRUE FALSE FALSE  TRUE
#> [13,] FALSE  TRUE FALSE FALSE FALSE
#> [14,] FALSE FALSE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE  TRUE  TRUE
#> [16,] FALSE  TRUE  TRUE  TRUE FALSE
#> [17,] FALSE  TRUE FALSE  TRUE FALSE
#> [18,] FALSE FALSE  TRUE FALSE  TRUE
#> [19,] FALSE FALSE FALSE FALSE FALSE
#> [20,] FALSE FALSE FALSE  TRUE  TRUE

A <= 10

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,] TRUE TRUE TRUE TRUE TRUE
#>  [2,] TRUE TRUE TRUE TRUE TRUE
#>  [3,] TRUE TRUE TRUE TRUE TRUE
#>  [4,] TRUE TRUE TRUE TRUE TRUE
#>  [5,] TRUE TRUE TRUE TRUE TRUE
#>  [6,] TRUE TRUE TRUE TRUE TRUE
#>  [7,] TRUE TRUE TRUE TRUE TRUE
#>  [8,] TRUE TRUE TRUE TRUE TRUE
#>  [9,] TRUE TRUE TRUE TRUE TRUE
#> [10,] TRUE TRUE TRUE TRUE TRUE
#> [11,] TRUE TRUE TRUE TRUE TRUE
#> [12,] TRUE TRUE TRUE TRUE TRUE
#> [13,] TRUE TRUE TRUE TRUE TRUE
#> [14,] TRUE TRUE TRUE TRUE TRUE
#> [15,] TRUE TRUE TRUE TRUE TRUE
#> [16,] TRUE TRUE TRUE TRUE TRUE
#> [17,] TRUE TRUE TRUE TRUE TRUE
#> [18,] TRUE TRUE TRUE TRUE TRUE
#> [19,] TRUE TRUE TRUE TRUE TRUE
#> [20,] TRUE TRUE TRUE TRUE TRUE

A >= 10

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE FALSE FALSE FALSE FALSE
#>  [2,] FALSE FALSE FALSE FALSE FALSE
#>  [3,] FALSE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE FALSE FALSE
#>  [6,] FALSE FALSE FALSE FALSE FALSE
#>  [7,] FALSE FALSE FALSE  TRUE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE
#>  [9,] FALSE  TRUE FALSE FALSE FALSE
#> [10,] FALSE  TRUE FALSE FALSE FALSE
#> [11,] FALSE FALSE FALSE  TRUE FALSE
#> [12,] FALSE FALSE FALSE FALSE FALSE
#> [13,] FALSE FALSE  TRUE FALSE FALSE
#> [14,] FALSE FALSE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE FALSE FALSE
#> [16,] FALSE FALSE FALSE FALSE FALSE
#> [17,] FALSE FALSE FALSE FALSE FALSE
#> [18,] FALSE FALSE FALSE  TRUE FALSE
#> [19,] FALSE FALSE FALSE  TRUE FALSE
#> [20,]  TRUE FALSE FALSE FALSE FALSE

## Arith
A + B

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]   17    1   13   11   11
#>  [2,]   12   10    3   11    2
#>  [3,]    3    6    9   18    9
#>  [4,]    5    7    4    0    0
#>  [5,]   11   15    4   11   11
#>  [6,]    8   11   15   12   11
#>  [7,]    2   13    3   16    8
#>  [8,]   14   15    5    3    8
#>  [9,]   14   19   12   17    5
#> [10,]    2   14    9    4   14
#> [11,]    9   16    8   18   16
#> [12,]   12    7   13    7    4
#> [13,]   11   12   19   17   16
#> [14,]    5   16    6   13   10
#> [15,]   12   17   13    4    2
#> [16,]   13    5   10   11   16
#> [17,]   16   11   14    4   11
#> [18,]    8    7    6   16    5
#> [19,]   11   12   15   14   19
#> [20,]   20   17    8    2    8

A * 2

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]   16    0   18    6   14
#>  [2,]   14   16    0    2    0
#>  [3,]    0   12   12   18   16
#>  [4,]    4    4    6    0    0
#>  [5,]   12   14    8    4   16
#>  [6,]    6   12   10    6   18
#>  [7,]    0   16    2   20    2
#>  [8,]   18   18    8    4   16
#>  [9,]    8   20    4   14    0
#> [10,]    0   20    6    4   10
#> [11,]    0   18    2   20   16
#> [12,]   18    4   18   14    2
#> [13,]   12    8   20   16   16
#> [14,]   10   12   12   10   18
#> [15,]   16   18   14    2    4
#> [16,]   14    2    6    2   18
#> [17,]   14    8   10    6   14
#> [18,]   16   14    6   20    6
#> [19,]   10   10   14   20   18
#> [20,]   20   16   12    0    0

A ^ 2

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]   64    0   81    9   49
#>  [2,]   49   64    0    1    0
#>  [3,]    0   36   36   81   64
#>  [4,]    4    4    9    0    0
#>  [5,]   36   49   16    4   64
#>  [6,]    9   36   25    9   81
#>  [7,]    0   64    1  100    1
#>  [8,]   81   81   16    4   64
#>  [9,]   16  100    4   49    0
#> [10,]    0  100    9    4   25
#> [11,]    0   81    1  100   64
#> [12,]   81    4   81   49    1
#> [13,]   36   16  100   64   64
#> [14,]   25   36   36   25   81
#> [15,]   64   81   49    1    4
#> [16,]   49    1    9    1   81
#> [17,]   49   16   25    9   49
#> [18,]   64   49    9  100    9
#> [19,]   25   25   49  100   81
#> [20,]  100   64   36    0    0

## Summary
max(A)

#> [1] 10

min(A)

#> [1] 0

range(A)

#> [1]  0 10

prod(A)

#> [1] 0

sum(A)

#> [1] 515

any(A)

#> [1] TRUE

all(A)

#> [1] FALSE

## Logic
A & B

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,]  TRUE FALSE  TRUE  TRUE  TRUE
#>  [2,]  TRUE  TRUE FALSE  TRUE FALSE
#>  [3,] FALSE FALSE  TRUE  TRUE  TRUE
#>  [4,]  TRUE  TRUE  TRUE FALSE FALSE
#>  [5,]  TRUE  TRUE FALSE  TRUE  TRUE
#>  [6,]  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [7,] FALSE  TRUE  TRUE  TRUE  TRUE
#>  [8,]  TRUE  TRUE  TRUE  TRUE FALSE
#>  [9,]  TRUE  TRUE  TRUE  TRUE FALSE
#> [10,] FALSE  TRUE  TRUE  TRUE  TRUE
#> [11,] FALSE  TRUE  TRUE  TRUE  TRUE
#> [12,]  TRUE  TRUE  TRUE FALSE  TRUE
#> [13,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [14,] FALSE  TRUE FALSE  TRUE  TRUE
#> [15,]  TRUE  TRUE  TRUE  TRUE FALSE
#> [16,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [17,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [18,] FALSE FALSE  TRUE  TRUE  TRUE
#> [19,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [20,]  TRUE  TRUE  TRUE FALSE FALSE

A | B

#>       [,1] [,2] [,3]  [,4]  [,5]
#>  [1,] TRUE TRUE TRUE  TRUE  TRUE
#>  [2,] TRUE TRUE TRUE  TRUE  TRUE
#>  [3,] TRUE TRUE TRUE  TRUE  TRUE
#>  [4,] TRUE TRUE TRUE FALSE FALSE
#>  [5,] TRUE TRUE TRUE  TRUE  TRUE
#>  [6,] TRUE TRUE TRUE  TRUE  TRUE
#>  [7,] TRUE TRUE TRUE  TRUE  TRUE
#>  [8,] TRUE TRUE TRUE  TRUE  TRUE
#>  [9,] TRUE TRUE TRUE  TRUE  TRUE
#> [10,] TRUE TRUE TRUE  TRUE  TRUE
#> [11,] TRUE TRUE TRUE  TRUE  TRUE
#> [12,] TRUE TRUE TRUE  TRUE  TRUE
#> [13,] TRUE TRUE TRUE  TRUE  TRUE
#> [14,] TRUE TRUE TRUE  TRUE  TRUE
#> [15,] TRUE TRUE TRUE  TRUE  TRUE
#> [16,] TRUE TRUE TRUE  TRUE  TRUE
#> [17,] TRUE TRUE TRUE  TRUE  TRUE
#> [18,] TRUE TRUE TRUE  TRUE  TRUE
#> [19,] TRUE TRUE TRUE  TRUE  TRUE
#> [20,] TRUE TRUE TRUE  TRUE  TRUE