Resample

Samples observations from a multinomial distribution.

Usage

resample(object, ...)

# S4 method for numeric
resample(object, do, n, ..., f = NULL)

Arguments

object

A numeric vector of count data (absolute frequencies).

...

Extra arguments passed to do.

do

A function that takes object as an argument and returns a single numeric value.

n

A non-negative integer giving the number of bootstrap replications.

f

A function that takes a single numeric vector (the result of do) as argument.

Value

If f is NULL, returns a numeric vector giving the n values of do. Else, returns the result of f applied to the n values of do.

See also

Author

N. Frerebeau

Examples

## Sample observations from a multinomial distribution
x <- sample(1:100, 50, TRUE) # integer
resample(x, do = median, n = 100)
#>   [1] 40.5 44.0 44.0 45.5 39.0 41.0 41.0 43.0 38.0 45.5 42.5 46.5 40.5 43.0 39.5
#>  [16] 43.0 39.0 41.0 37.0 36.5 41.0 46.0 43.0 41.0 42.0 42.5 39.5 41.0 42.5 44.5
#>  [31] 43.5 43.0 41.0 43.0 42.0 39.0 42.0 43.0 43.5 43.0 47.5 44.0 40.5 40.5 46.0
#>  [46] 43.5 40.5 39.5 44.5 35.0 46.0 39.0 38.5 40.5 42.0 39.0 41.5 41.0 38.5 42.0
#>  [61] 39.5 46.0 38.5 42.0 40.5 43.5 45.0 42.5 43.0 40.5 43.0 39.0 44.0 40.5 42.0
#>  [76] 40.0 44.0 41.5 42.0 43.0 39.5 42.5 44.5 42.5 39.5 42.5 40.5 38.0 39.0 39.5
#>  [91] 40.0 41.0 39.5 41.5 42.0 40.0 45.5 40.5 41.5 46.0

## Estimate the 25th, 50th and 95th percentiles
quant <- function(x) { quantile(x, probs = c(0.25, 0.50, 0.75)) }
resample(x, n = 100, do = median, f = quant)
#>  25%  50%  75% 
#> 40.0 41.5 43.0 

## Jackknife
x <- rnorm(20)
jackknife(x, do = mean) # Sample mean
#>       mean       bias      error 
#> 0.02074943 0.00000000 0.20474580