Absolute Frequency Matrix

An S4 class to represent an absolute frequency matrix (i.e. the number of times a given datum occurs in a dataset).

Usage

CountMatrix(data = 0, nrow = 1, ncol = 1, byrow = FALSE, dimnames = NULL)

Arguments

data

an optional data vector (including a list or expression vector). Non-atomic classed R objects are coerced by as.vector and all attributes discarded.

nrow

the desired number of rows.

ncol

the desired number of columns.

byrow

logical. If FALSE (the default) the matrix is filled by columns, otherwise the matrix is filled by rows.

dimnames

A dimnames attribute for the matrix: NULL or a list of length 2 giving the row and column names respectively. An empty list is treated as NULL, and a list of length one as row names. The list can be named, and the list names will be used as names for the dimensions.

See also

Author

N. Frerebeau

Examples

## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)

## Access
dim(B) # Get the matrix dimensions
#> [1] 20  5
row(B) # Get the row indexes
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1    1    1    1    1
#>  [2,]    2    2    2    2    2
#>  [3,]    3    3    3    3    3
#>  [4,]    4    4    4    4    4
#>  [5,]    5    5    5    5    5
#>  [6,]    6    6    6    6    6
#>  [7,]    7    7    7    7    7
#>  [8,]    8    8    8    8    8
#>  [9,]    9    9    9    9    9
#> [10,]   10   10   10   10   10
#> [11,]   11   11   11   11   11
#> [12,]   12   12   12   12   12
#> [13,]   13   13   13   13   13
#> [14,]   14   14   14   14   14
#> [15,]   15   15   15   15   15
#> [16,]   16   16   16   16   16
#> [17,]   17   17   17   17   17
#> [18,]   18   18   18   18   18
#> [19,]   19   19   19   19   19
#> [20,]   20   20   20   20   20
col(B, as.factor = TRUE) # Get the column indexes
#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,] col1 col2 col3 col4 col5
#>  [2,] col1 col2 col3 col4 col5
#>  [3,] col1 col2 col3 col4 col5
#>  [4,] col1 col2 col3 col4 col5
#>  [5,] col1 col2 col3 col4 col5
#>  [6,] col1 col2 col3 col4 col5
#>  [7,] col1 col2 col3 col4 col5
#>  [8,] col1 col2 col3 col4 col5
#>  [9,] col1 col2 col3 col4 col5
#> [10,] col1 col2 col3 col4 col5
#> [11,] col1 col2 col3 col4 col5
#> [12,] col1 col2 col3 col4 col5
#> [13,] col1 col2 col3 col4 col5
#> [14,] col1 col2 col3 col4 col5
#> [15,] col1 col2 col3 col4 col5
#> [16,] col1 col2 col3 col4 col5
#> [17,] col1 col2 col3 col4 col5
#> [18,] col1 col2 col3 col4 col5
#> [19,] col1 col2 col3 col4 col5
#> [20,] col1 col2 col3 col4 col5
#> Levels: col1 col2 col3 col4 col5
nrow(B) # Get the number of rows
#> [1] 20
ncol(B) # Get the number of columns
#> [1] 5
dimnames(B) # Get the dimension names
#> [[1]]
#>  [1] "row1"  "row2"  "row3"  "row4"  "row5"  "row6"  "row7"  "row8"  "row9" 
#> [10] "row10" "row11" "row12" "row13" "row14" "row15" "row16" "row17" "row18"
#> [19] "row19" "row20"
#> 
#> [[2]]
#> [1] "col1" "col2" "col3" "col4" "col5"
#> 
rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T"
colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names
#> [1] "u" "v" "w" "x" "y"

## Subset
B[[1, 1]] # Get the first value
#> [1] 4
B[1] # Get the first value
#> [1] 4
B[, ] # Get all values
#> <CountMatrix: 20 x 5>
#>    u  v  w  x  y
#> A  4  9  7  1  9
#> B  8  9  9 10  8
#> C 10  3  4  5  5
#> D  1  5 10  3 10
#> E  4  8  8  9 10
#> F 10 10 10  6  1
#> G  2 10  0  2  2
#> H  2  9  5  1  9
#> I  9 10  6  3  6
#> J  9  5  8  1  9
#> K  2 10  6  5  2
#> L  3  9  9  9  0
#> M  6  2  8  6  7
#> N 10  7  8  2  4
#> O  0  0 10  4  2
#> P  1 10  1  5  2
#> Q  0  5  5  3 10
#> R  9  3  8  9  0
#> S  1  7  2  3 10
#> T  3  4  5  2  4
B[1, , drop = FALSE] # Get the first row
#> <CountMatrix: 1 x 5>
#>   u v w x y
#> A 4 9 7 1 9
B[, 1:3] # Get the first three column
#> <CountMatrix: 20 x 3>
#>    u  v  w
#> A  4  9  7
#> B  8  9  9
#> C 10  3  4
#> D  1  5 10
#> E  4  8  8
#> F 10 10 10
#> G  2 10  0
#> H  2  9  5
#> I  9 10  6
#> J  9  5  8
#> K  2 10  6
#> L  3  9  9
#> M  6  2  8
#> N 10  7  8
#> O  0  0 10
#> P  1 10  1
#> Q  0  5  5
#> R  9  3  8
#> S  1  7  2
#> T  3  4  5