Bootstrap Estimation

Creates \(n\) bootstrapped replicates by resampling with replacement from the original data and applies a function.

bootstrap(x, ...)

# S4 method for numeric
bootstrap(x, do, n, ...)

# S4 method for integer
bootstrap(x, do, n, ...)

# S4 method for BootstrapVector
summary(
  object,
  level = 0.95,
  type = c("student", "normal"),
  probs = c(0.25, 0.75),
  na.rm = FALSE,
  ...
)

Arguments

x

A numeric or an integer vector (see below).

...

Extra arguments passed to do.

do

A function that takes x as an argument and returns a single numeric value.

n

A non-negative integer giving the number of bootstrap replications.

object

A BootstrapVector object.

level

A length-one numeric vector giving the confidence level. Must be a single number between \(0\) and \(1\). If NULL, no confidence interval are computed.

type

A character string giving the type of confidence interval to be returned. It must be one "student" (default) or "normal". Any unambiguous substring can be given. Only used if level is not NULL.``

probs

A numeric vector of probabilities with values in \([0,1]\) (see stats::quantile()). If NULL, quantiles are not computed.

na.rm

A logical scalar: should missing values be removed from object before the sample statistics are computed?

Value

bootstrap() returns a BootstrapVector object (i.e. a numeric vector of the n bootstrap values of do).

summary() returns a numeric vector with the following elements:

min

Minimum value.

mean

Mean value.

max

Maximum value.

lower

Lower bound of the confidence interval.

upper

Upper bound of the confidence interval.

Q*

Sample quantile to * probability.

Methods (by class)

numeric: Samples randomly from the elements of x with replacement.

integer: Samples observations from a multinomial distribution.

See also

Author

N. Frerebeau

Examples

## Jackknife
x <- rnorm(20)
jack <- jackknife(x, do = mean) # Sample mean
summary(jack)
#>         mean         bias        error 
#> -0.006986036  0.000000000  0.219341693 

## Random samples from x with replacement
x <- rnorm(20) # numeric
boot <- bootstrap(x, do = mean, n = 100) # Sample mean
summary(boot)
#>         min        mean         max       lower       upper         Q25 
#> -0.44037777  0.13172617  0.60802952  0.09123151  0.17222083 -0.01904940 
#>         Q75 
#>  0.25660783 

## Sample observations from a multinomial distribution
x <- sample(1:100, 100, TRUE) # integer
boot <- bootstrap(x, do = median, n = 100)
summary(boot)
#>      min     mean      max    lower    upper      Q25      Q75 
#> 53.00000 57.20500 61.00000 56.85698 57.55302 56.00000 58.50000