Bootstrap Estimation

Samples randomly from the elements of object with replacement.

Usage

bootstrap(object, ...)

# S4 method for numeric
bootstrap(object, do, n, ..., f = NULL)

Arguments

object

A numeric vector.

...

Extra arguments to be passed to do.

do

A function that takes object as an argument and returns a single numeric value.

n

A non-negative integer giving the number of bootstrap replications.

f

A function that takes a single numeric vector (the result of do) as argument.

Value

If f is NULL (the default), bootstrap() returns a named numeric

vector with the following elements:

original

The observed value of do applied to object.

mean

The bootstrap estimate of mean of do.

bias

The bootstrap estimate of bias of do.

error

he bootstrap estimate of standard error of do.

If f is a function, bootstrap() returns the result of f applied to the n values of do.

See also

Author

N. Frerebeau

Examples

x <- rnorm(20)

## Bootstrap
bootstrap(x, do = mean, n = 100)
#>    original        mean        bias       error 
#> -0.15047904 -0.16077539 -0.01029636  0.17865800 

## Estimate the 25th and 95th percentiles
quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
bootstrap(x, n = 100, do = mean, f = quant)
#>         25%         75% 
#> -0.23204625  0.01022648 

## Get the n bootstrap values
bootstrap(x, n = 100, do = mean, f = function(x) { x })
#>   [1] -0.046736312 -0.179189812  0.208986661 -0.303983091 -0.220330711
#>   [6] -0.023110520  0.051241680  0.272642502  0.020461214 -0.601131162
#>  [11] -0.546780890 -0.549871079 -0.149642336 -0.202370568  0.260250922
#>  [16]  0.150367024 -0.219366415 -0.052001932 -0.169804755 -0.389532973
#>  [21] -0.084813267 -0.195865748  0.143500975 -0.160755369 -0.116810989
#>  [26] -0.129129130  0.021505616  0.262129804 -0.232037214 -0.302534903
#>  [31]  0.154929849 -0.001771162  0.162122515  0.111923358 -0.152821893
#>  [36]  0.333027412  0.116412370 -0.006880780 -0.204308570  0.094005166
#>  [41] -0.037489437 -0.078380794 -0.168384542 -0.104014059 -0.092234686
#>  [46] -0.354511527 -0.291788214 -0.355178198 -0.134430771  0.070828889
#>  [51] -0.226790025 -0.557920324  0.119007998 -0.146926943 -0.107744076
#>  [56] -0.037084823 -0.095503345 -0.108057635  0.035976539 -0.153587539
#>  [61] -0.242821199 -0.163843504 -0.087884079 -0.293884437 -0.423566954
#>  [66] -0.049190007  0.058421544 -0.240700075 -0.166346647  0.043514133
#>  [71]  0.002915850 -0.490500184 -0.054755273 -0.059324698 -0.377040990
#>  [76]  0.236806277 -0.178756251 -0.134848998 -0.000730129 -0.166614999
#>  [81]  0.086317965 -0.327876402 -0.153269875 -0.311683642  0.089288461
#>  [86]  0.280205547  0.083743900  0.040272469 -0.260622329 -0.050520605
#>  [91] -0.487340940 -0.179168469  0.173098497 -0.017911805  0.292574789
#>  [96] -0.064228632 -0.012385358 -0.544512250 -0.110325227 -0.442699849

## Jackknife
jackknife(x, do = mean) # Sample mean
#>   original       mean       bias      error 
#> -0.1504790 -0.1504790  0.0000000  0.2094266 

## Get the leave-one-out values instead of summary
jackknife(x, do = mean, f = function(x) { x })
#>  [1] -0.07398475 -0.19608310 -0.07467077 -0.12474718 -0.11916590 -0.14865509
#>  [7] -0.16781654 -0.15886272 -0.15585347 -0.14620002 -0.16685928 -0.11341496
#> [13] -0.22215653 -0.08364132 -0.24632480 -0.13640547 -0.16080426 -0.10061924
#> [19] -0.22774928 -0.18556606