Correlation, Variance and Covariance

Computes the variance, covariance or correlation of a DataMatrix.

# S4 method for DataMatrix,missing
var(x, na.rm = FALSE)

# S4 method for DataMatrix,missing
cov(x, use = "everything", method = c("pearson", "kendall", "spearman"))

# S4 method for DataMatrix,missing
cor(x, use = "everything", method = c("pearson", "kendall", "spearman"))

Arguments

x
A DataMatrix object.

na.rm
A logical scalar: should missing values (including NaN) be omitted from the calculations?

use
An optional character string giving a method for computing covariances in the presence of missing values (see cor).

method
A character string indicating which correlation coefficient (or covariance) is to be computed (see cor).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column means
rowMeans(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>   7.6   6.0   4.0   5.2   6.6   5.6   5.4   5.6   4.4   5.2   5.6   3.6   5.6 
#> row14 row15 row16 row17 row18 row19 row20 
#>   6.2   6.8   5.2   7.4   5.8   3.6   6.0 

colMeans(A)

#> col1 col2 col3 col4 col5 
#> 5.85 6.05 5.00 5.30 5.65 

## Row and column sums
rowSums(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>    38    30    20    26    33    28    27    28    22    26    28    18    28 
#> row14 row15 row16 row17 row18 row19 row20 
#>    31    34    26    37    29    18    30 

colSums(A)

#> col1 col2 col3 col4 col5 
#>  117  121  100  106  113 

## Variance, covariance and correlation matrix
var(A)

#>           col1       col2      col3       col4       col5
#> col1  7.713158 -3.9394737 -1.736842  1.5736842 -1.0026316
#> col2 -3.939474  8.0500000  1.578947 -0.8578947  0.2815789
#> col3 -1.736842  1.5789474  9.894737 -1.6842105  1.6315789
#> col4  1.573684 -0.8578947 -1.684211  8.6421053 -2.9421053
#> col5 -1.002632  0.2815789  1.631579 -2.9421053 10.0289474

cov(A)

#>           col1       col2      col3       col4       col5
#> col1  7.713158 -3.9394737 -1.736842  1.5736842 -1.0026316
#> col2 -3.939474  8.0500000  1.578947 -0.8578947  0.2815789
#> col3 -1.736842  1.5789474  9.894737 -1.6842105  1.6315789
#> col4  1.573684 -0.8578947 -1.684211  8.6421053 -2.9421053
#> col5 -1.002632  0.2815789  1.631579 -2.9421053 10.0289474

cor(A)

#>            col1        col2       col3       col4        col5
#> col1  1.0000000 -0.49994721 -0.1988118  0.1927488 -0.11399809
#> col2 -0.4999472  1.00000000  0.1769162 -0.1028552  0.03133824
#> col3 -0.1988118  0.17691619  1.0000000 -0.1821313  0.16378657
#> col4  0.1927488 -0.10285521 -0.1821313  1.0000000 -0.31602450
#> col5 -0.1139981  0.03133824  0.1637866 -0.3160245  1.00000000