Correlation, Variance and Covariance

Computes the variance, covariance or correlation of a DataMatrix.

# S4 method for DataMatrix,missing
var(x, na.rm = FALSE)

# S4 method for DataMatrix,missing
cov(x, use = "everything", method = c("pearson", "kendall", "spearman"))

# S4 method for DataMatrix,missing
cor(x, use = "everything", method = c("pearson", "kendall", "spearman"))

Arguments

x
A DataMatrix object.

na.rm
A logical scalar: should missing values (including NaN) be omitted from the calculations?

use
An optional character string giving a method for computing covariances in the presence of missing values (see cor).

method
A character string indicating which correlation coefficient (or covariance) is to be computed (see cor).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Row and column means
rowMeans(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>   4.4   5.0   4.8   5.8   6.8   4.6   4.8   4.6   5.0   6.0   5.8   5.8   5.4 
#> row14 row15 row16 row17 row18 row19 row20 
#>   7.0   7.2   3.4   6.0   7.2   7.2   3.4 

colMeans(A)

#> col1 col2 col3 col4 col5 
#> 4.75 5.60 5.50 6.40 5.30 

## Row and column sums
rowSums(A)

#>  row1  row2  row3  row4  row5  row6  row7  row8  row9 row10 row11 row12 row13 
#>    22    25    24    29    34    23    24    23    25    30    29    29    27 
#> row14 row15 row16 row17 row18 row19 row20 
#>    35    36    17    30    36    36    17 

colSums(A)

#> col1 col2 col3 col4 col5 
#>   95  112  110  128  106 

## Variance, covariance and correlation matrix
var(A)

#>            col1        col2       col3       col4        col5
#> col1  8.1973684 -1.26315789  0.2368421 -1.1578947 -1.76315789
#> col2 -1.2631579 10.25263158 -1.5789474  1.3789474 -0.03157895
#> col3  0.2368421 -1.57894737  9.6315789  1.4736842  0.36842105
#> col4 -1.1578947  1.37894737  1.4736842  4.8842105 -0.75789474
#> col5 -1.7631579 -0.03157895  0.3684211 -0.7578947  8.11578947

cov(A)

#>            col1        col2       col3       col4        col5
#> col1  8.1973684 -1.26315789  0.2368421 -1.1578947 -1.76315789
#> col2 -1.2631579 10.25263158 -1.5789474  1.3789474 -0.03157895
#> col3  0.2368421 -1.57894737  9.6315789  1.4736842  0.36842105
#> col4 -1.1578947  1.37894737  1.4736842  4.8842105 -0.75789474
#> col5 -1.7631579 -0.03157895  0.3684211 -0.7578947  8.11578947

cor(A)

#>             col1         col2        col3       col4         col5
#> col1  1.00000000 -0.137785253  0.02665465 -0.1829931 -0.216166600
#> col2 -0.13778525  1.000000000 -0.15889171  0.1948645 -0.003461901
#> col3  0.02665465 -0.158891708  1.00000000  0.2148614  0.041670653
#> col4 -0.18299308  0.194864545  0.21486145  1.0000000 -0.120377776
#> col5 -0.21616660 -0.003461901  0.04167065 -0.1203778  1.000000000