Correspondence Analysis

Computes a simple correspondence analysis based on the singular value decomposition.

ca(object, ...)

get_coordinates(object, ...)

get_contributions(object, ...)

get_distances(object, ...)

get_eigenvalues(object)

get_inertia(object, ...)

# S4 method for CountMatrix
ca(object, n = min(dim(object) - 1), sup_rows = NULL, sup_columns = NULL)

# S4 method for CA
predict(object, data, margin = 1)

# S4 method for CA
get_coordinates(
  object,
  margin = 1,
  standard = FALSE,
  sup = TRUE,
  sup_name = ".sup"
)

# S4 method for CA
get_inertia(object, margin = 1)

# S4 method for CA
get_distances(object, margin = 1)

# S4 method for CA
get_contributions(object, margin = 1)

# S4 method for CA
get_eigenvalues(object)

Arguments

object
A CountMatrix or a CA object.

...
Currently not used.

n
An integer value specifying the number of dimensions to be kept in the results (default to 5).

sup_rows
A numeric or logical vector specifying the indices of the supplementary rows.

sup_columns
A numeric or logical vector specifying the indices of the supplementary columns.

data
An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates rows (the default), 2 indicates columns.

standard
A logical scalar: should standard coordinates be returned instead of principal coordinates (the default)?

sup
A logical scalar: should supplementary points be returned?

sup_name
A character string specifying the name of the column to create for supplementary points attribution (see below).

Value

ca returns a CA object.

predict returns a data.frame of principal coordinates.

get_coordinates returns a data.frame of coordinates. If sup is TRUE, an extra column (named after sup_name) is added specifying whether an observation is a supplementary point or not.

get_inertia returns a data.frame.

get_distances returns a data.frame.

get_contributions returns a data.frame.

get_eigenvalues returns a data.frame with the following columns: eigenvalues, percentage (percentage of variance) and cumulative (cumulative percentage of variance).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- CountMatrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Predict new row coordinates
predict(X, B, margin = 1)

#>               CA1          CA2          CA3        CA4           CA5
#> row1 -0.007820561 -0.090490566 -0.007455178 0.10919821  0.0009590692
#> row2 -0.215506980 -0.041050657 -0.012922626 0.02873122 -0.0487594129
#> row3 -0.101011488  0.008160103  0.059542122 0.22309978  0.0991063370
#> row4 -0.060603459 -0.078551810  0.014002693 0.09741182 -0.0396211920
#> row5 -0.131123926  0.079345693 -0.080528794 0.06485535  0.0559515390
#> row6  0.006139721 -0.082332451  0.034390111 0.05497076  0.1443931073
#> row7 -0.177653381  0.060753306 -0.073761414 0.08191048  0.0533543084

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

#>                 CA1          CA2           CA3         CA4          CA5  .sup
#> row1  -0.0005820818  0.148394896 -0.2424179912 -0.07285316  0.002794611 FALSE
#> row2  -0.1025326962 -0.492290839 -0.0419879806 -0.02278684  0.027276041 FALSE
#> row3   0.2537238320 -0.111370511  0.0001890362  0.10206364 -0.004069772 FALSE
#> row4   0.4547949047 -0.003187552 -0.0317495436 -0.02701901 -0.027374085 FALSE
#> row5  -0.8866005980  0.155714766 -0.1005719584  0.05376078 -0.013059829 FALSE
#> row6   0.2068363684  0.444924973  0.1479724189  0.01321923  0.031994786 FALSE
#> row7  -0.2847068282 -0.081904440  0.3931354055 -0.06438232 -0.014341846 FALSE
#> row8   0.0771909194  0.101443929 -0.0249884596 -0.27053595 -0.279457477  TRUE
#> row9   0.1411873700  0.133236617  0.0094642916  0.42731515  0.183248323  TRUE
#> row10 -0.5593496194 -0.045845589 -0.3349975624 -0.19790245 -0.155992419  TRUE

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

#>              CA1         CA2           CA3         CA4           CA5  .sup
#> col1   0.1280548 -0.18180310  8.910881e-02 -0.05992428 -0.0283656300 FALSE
#> col2   0.3219454  0.01207608  8.975048e-05 -0.06592359  0.0323586478 FALSE
#> col3  -0.5328685  0.27810042  2.585591e-01  0.01529428  0.0043559177 FALSE
#> col4  -0.6106188 -0.10451705 -3.205783e-01 -0.01414365  0.0009023154 FALSE
#> col5   0.1279605 -0.41046012  6.943129e-02  0.11099659  0.0091322039 FALSE
#> col6   0.4093789  0.37795996 -1.408004e-01  0.05237479 -0.0127182629 FALSE
#> col7  -0.2301894 -0.12400165  8.420324e-02 -0.30372194 -0.0997613194  TRUE
#> col8   0.1258193  0.09918414  7.788190e-02  0.05853030  0.0645330146  TRUE
#> col9   0.4521520 -0.00210787  1.641146e-01  0.13553497  0.0289811417  TRUE
#> col10 -0.3037318  0.06146369  2.500790e-02 -0.08945515 -0.1101622013  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>      row1      row2      row3      row4      row5      row6      row7      row8 
#> 0.2934336 0.5058552 0.2953181 0.4575325 0.9074598 0.5136475 0.4966614 0.4100731 
#>      row9     row10 
#> 0.5039381 0.7004971 

## Get row inertias
get_inertia(X, margin = 1)

#>       row1       row2       row3       row4       row5       row6       row7 
#> 0.01405767 0.03551119 0.01601867 0.03417730 0.09411237 0.03338305 0.02718432 
#>       row8       row9      row10 
#>         NA         NA         NA 

## Get row contributions
get_contributions(X, margin = 1)

#>               CA1          CA2          CA3        CA4        CA5
#> row1 3.657732e-05  5.282074263 3.098720e+01 22.9223790  0.3175069
#> row2 9.646894e-01 49.411610396 7.901718e-01  1.9061185 25.7094115
#> row3 7.818418e+00  3.347029356 2.119802e-05 50.6124724  0.7575358
#> row4 2.232930e+01  0.002437143 5.315298e-01  3.1528385 30.4641995
#> row5 5.940157e+01  4.071216598 3.733400e+00  8.7375981  4.8538187
#> row6 3.579310e+00 36.799493677 8.947792e+00  0.5848936 32.2530409
#> row7 5.906679e+00  1.086138567 5.500988e+01 12.0837000  5.6444866

## Get eigenvalues
get_eigenvalues(X)

#>      eigenvalues percentage cumulative
#> CA1 0.1512342163 59.4369994   59.43700
#> CA2 0.0680655460 26.7506383   86.18764
#> CA3 0.0309628737 12.1688091   98.35645
#> CA4 0.0037803413  1.4857229   99.84217
#> CA5 0.0004015904  0.1578302  100.00000