Jackknife Estimation

Jackknife Estimation

Usage

jackknife(object, ...)

# S4 method for numeric
jackknife(object, do, ..., f = NULL)

Arguments

object

A numeric vector.

...

Extra arguments to be passed to do.

do

A function that takes object as an argument and returns a single numeric value.

f

A function that takes a single numeric vector (the leave-one-out values of do) as argument.

Value

If f is NULL (the default), jackknife() returns a named numeric

vector with the following elements:

original

The observed value of do applied to object.

mean

The jackknife estimate of mean of do.

bias

The jackknife estimate of bias of do.

error

he jackknife estimate of standard error of do.

If f is a function, jackknife() returns the result of f applied to the leave-one-out values of do.

See also

Author

N. Frerebeau

Examples

x <- rnorm(20)

## Bootstrap
bootstrap(x, do = mean, n = 100)
#>    original        mean        bias       error 
#> -0.17119796 -0.21238609 -0.04118813  0.22981870 

## Estimate the 25th and 95th percentiles
quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
bootstrap(x, n = 100, do = mean, f = quant)
#>         25%         75% 
#> -0.36220980 -0.04700279 

## Get the n bootstrap values
bootstrap(x, n = 100, do = mean, f = function(x) { x })
#>   [1] -0.065575074  0.262082539 -0.107008116 -0.155062480 -0.281937348
#>   [6]  0.274118683 -0.221600293 -0.017325902 -0.601938938 -0.191279345
#>  [11]  0.332193141  0.044913512 -0.112748557 -0.289217040 -0.465629550
#>  [16]  0.155026426 -0.783901193 -0.500476361 -0.176277774 -0.359933492
#>  [21] -0.254769393 -0.411230565 -0.090432456  0.155402802  0.017031990
#>  [26] -0.234840769 -0.432971641  0.086058201 -0.202570785 -0.128901562
#>  [31] -0.126297740 -0.135299574 -0.447745899 -0.178852730  0.523505471
#>  [36]  0.106361636 -0.123388751 -0.294037314 -0.340683843 -0.376602003
#>  [41]  0.410550113 -0.284350085 -0.323318080 -0.087522266 -0.181670911
#>  [46] -0.305275146 -0.029657265  0.193871716 -0.553949638  0.016640376
#>  [51] -0.513730932 -0.002247111 -0.714556787 -0.525033366 -0.250442795
#>  [56] -0.449260067 -0.338824820 -0.279417244 -0.029686009  0.028549143
#>  [61] -0.929947793 -0.248394106 -0.770846406  0.108236967 -0.329278578
#>  [66]  0.036614360 -0.439271746 -0.484519117 -0.385852956 -0.451153048
#>  [71] -0.275842420 -0.277768103  0.035018368 -0.066876119  0.181378736
#>  [76] -0.363785369 -0.126049761 -0.342181737  0.012285295 -0.465431800
#>  [81] -0.001388738 -0.353174567 -0.171640375 -0.161999381 -0.280640119
#>  [86] -0.092119736 -0.296014660  0.077730456 -0.136019604 -0.068566055
#>  [91] -0.576109840  0.060538862 -0.510064664  0.013830669 -0.210373174
#>  [96]  0.080965162 -0.322095963 -0.096570738  0.212741780 -0.247312835

## Jackknife
jackknife(x, do = mean) # Sample mean
#>  original      mean      bias     error 
#> -0.171198 -0.171198  0.000000  0.254635 

## Get the leave-one-out values instead of summary
jackknife(x, do = mean, f = function(x) { x })
#>  [1] -0.15999965 -0.24544348 -0.06553019 -0.20219098 -0.30504448 -0.07868733
#>  [7] -0.14664390 -0.07862861 -0.24082488 -0.13044111 -0.24402689 -0.13963148
#> [13] -0.15336333 -0.19850297 -0.17973442 -0.16616054 -0.17984629 -0.15120178
#> [19] -0.16732959 -0.19072721