Get or Set Parts of an Object

Getters and setters to retrieve or set parts of an object.

get_id(x)

set_id(x)

set_id(x) <- value

get_method(x)

set_method(x) <- value

get_units(x)

get_totals(x)

set_totals(x) <- value

# S4 method for DataMatrix
length(x)

# S4 method for DataMatrix
diag(x)

# S4 method for DataMatrix
dim(x)

# S4 method for DataMatrix
row(x, as.factor = FALSE)

# S4 method for DataMatrix
col(x, as.factor = FALSE)

# S4 method for DataMatrix
nrow(x)

# S4 method for DataMatrix
ncol(x)

# S4 method for DataMatrix
dimnames(x)

# S4 method for DataMatrix
dimnames(x) <- value

# S4 method for DataMatrix
rownames(x)

# S4 method for DataMatrix
rownames(x) <- value

# S4 method for DataMatrix
colnames(x)

# S4 method for DataMatrix
colnames(x) <- value

# S4 method for ANY
get_id(x)

# S4 method for AbundanceMatrix
get_totals(x)

# S4 method for SimilarityMatrix
get_method(x)

# S4 method for StratigraphicMatrix
get_units(x)

# S4 method for ANY
set_id(x)

# S4 method for ANY
set_id(x) <- value

# S4 method for AbundanceMatrix
set_totals(x) <- value

# S4 method for SimilarityMatrix
set_method(x) <- value

Arguments

x
An object from which to get or set element(s).

value
A possible value for the element(s) of object (see below).

as.factor
A logical scalar: should the value be returned as a factor of row or column labels rather than as numbers?

Value

An object of the same sort as x with the new values assigned.

Functions

length,DataMatrix-method: Retrieve the length of an object.

diag,DataMatrix-method: Extract the diagonal of a matrix.

dim,DataMatrix-method: Retrieve or set the dimension of an object.

row,DataMatrix-method: Returns a matrix of integers indicating their row number in a matrix-like object, or a factor indicating the row labels.

col,DataMatrix-method: Returns a matrix of integers indicating their column number in a matrix-like object, or a factor of column labels.

nrow,DataMatrix-method: Return the number of rows present in x.

ncol,DataMatrix-method: Return the number of columns present in x.

dimnames,DataMatrix-method: Retrieve the dimnames of x.

dimnames<-,DataMatrix-method: Set the dimnames of x.

rownames,DataMatrix-method: Retrieve the row names of x.

rownames<-,DataMatrix-method: Set the row names of x.

colnames,DataMatrix-method: Retrieve the column names of x.

colnames<-,DataMatrix-method: Set the column names of x.

See also

Examples

## Create an incidence (presence/absence) matrix
## Data will be coerced with as.logical()
A <- IncidenceMatrix(data = sample(0:1, 100, TRUE, c(1, 1/3)), nrow = 20)
## Create a count data matrix
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20)
get_id(B)

#> [1] "89e31a74-7dae-4a3f-a711-ac8bbed3cd14"

## Access
dim(B) # Get the matrix dimensions

#> [1] 20  5

row(B) # Get the row indexes

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    1    1    1    1    1
#>  [2,]    2    2    2    2    2
#>  [3,]    3    3    3    3    3
#>  [4,]    4    4    4    4    4
#>  [5,]    5    5    5    5    5
#>  [6,]    6    6    6    6    6
#>  [7,]    7    7    7    7    7
#>  [8,]    8    8    8    8    8
#>  [9,]    9    9    9    9    9
#> [10,]   10   10   10   10   10
#> [11,]   11   11   11   11   11
#> [12,]   12   12   12   12   12
#> [13,]   13   13   13   13   13
#> [14,]   14   14   14   14   14
#> [15,]   15   15   15   15   15
#> [16,]   16   16   16   16   16
#> [17,]   17   17   17   17   17
#> [18,]   18   18   18   18   18
#> [19,]   19   19   19   19   19
#> [20,]   20   20   20   20   20

col(B, as.factor = TRUE) # Get the column indexes

#>       [,1]   [,2]   [,3]   [,4]   [,5]  
#>  [1,] "col1" "col2" "col3" "col4" "col5"
#>  [2,] "col1" "col2" "col3" "col4" "col5"
#>  [3,] "col1" "col2" "col3" "col4" "col5"
#>  [4,] "col1" "col2" "col3" "col4" "col5"
#>  [5,] "col1" "col2" "col3" "col4" "col5"
#>  [6,] "col1" "col2" "col3" "col4" "col5"
#>  [7,] "col1" "col2" "col3" "col4" "col5"
#>  [8,] "col1" "col2" "col3" "col4" "col5"
#>  [9,] "col1" "col2" "col3" "col4" "col5"
#> [10,] "col1" "col2" "col3" "col4" "col5"
#> [11,] "col1" "col2" "col3" "col4" "col5"
#> [12,] "col1" "col2" "col3" "col4" "col5"
#> [13,] "col1" "col2" "col3" "col4" "col5"
#> [14,] "col1" "col2" "col3" "col4" "col5"
#> [15,] "col1" "col2" "col3" "col4" "col5"
#> [16,] "col1" "col2" "col3" "col4" "col5"
#> [17,] "col1" "col2" "col3" "col4" "col5"
#> [18,] "col1" "col2" "col3" "col4" "col5"
#> [19,] "col1" "col2" "col3" "col4" "col5"
#> [20,] "col1" "col2" "col3" "col4" "col5"

nrow(B) # Get the number of rows

#> [1] 20

ncol(B) # Get the number of columns

#> [1] 5

dimnames(B) # Get the dimension names

#> [[1]]
#>  [1] "row1"  "row2"  "row3"  "row4"  "row5"  "row6"  "row7"  "row8"  "row9" 
#> [10] "row10" "row11" "row12" "row13" "row14" "row15" "row16" "row17" "row18"
#> [19] "row19" "row20"
#> 
#> [[2]]
#> [1] "col1" "col2" "col3" "col4" "col5"
#> 

rownames(B) <- LETTERS[1:20] # Set the row names
rownames(B) # Get the rownames

#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T"

colnames(B) <- letters[21:25] # Set the column names
colnames(B) # Get the column names

#> [1] "u" "v" "w" "x" "y"

## Subset
B[[1, 1]] # Get the first value

#> [1] 0

B[, ] # Get all values

#>   u  v  w  x  y
#> A 0  5  2  3 10
#> B 8  3  0  3  9
#> C 0  3  5  1  0
#> D 8  5  8  0  9
#> E 9  5  5 10  9
#> F 1  2  5  2  6
#> G 8  5  6  1  1
#> H 0 10  9 10 10
#> I 5  2  4  6  2
#> J 8  9  7  7  8
#> K 1  3  5  4  0
#> L 8  5  8  9  9
#> M 9  0 10  0  8
#> N 2  4  8  6  3
#> O 9  6  4  7 10
#> P 7  9  7  9  1
#> Q 3  0  0  3  6
#> R 9  6  7  8  4
#> S 0 10  9  2  2
#> T 9  1  4  8  4

B[1, ] # Get the first row

#>  u  v  w  x  y 
#>  0  5  2  3 10 

B[, 1] # Get the first column

#> A B C D E F G H I J K L M N O P Q R S T 
#> 0 8 0 8 9 1 8 0 5 8 1 8 9 2 9 7 3 9 0 9 

B[c("A", "B", "C"), ] # Get the first three rows

#>   u v w x  y
#> A 0 5 2 3 10
#> B 8 3 0 3  9
#> C 0 3 5 1  0

B[c("A", "B", "C"), 1, drop = FALSE] # Get the first 3 rows of the first column

#>   u
#> A 0
#> B 8
#> C 0