Common Operations on Matrix Objects

Performs common operations on DataMatrix objects.

# S4 method for DataMatrix,DataMatrix
Arith(e1, e2)

# S4 method for DataMatrix,numeric
Arith(e1, e2)

# S4 method for DataMatrix,DataMatrix
Compare(e1, e2)

# S4 method for DataMatrix,numeric
Compare(e1, e2)

# S4 method for DataMatrix,DataMatrix
Logic(e1, e2)

# S4 method for DataMatrix,numeric
Logic(e1, e2)

# S4 method for DataMatrix
Summary(x, na.rm = FALSE)

# S4 method for DataMatrix
rowAll(x, f, ..., na.rm = FALSE)

# S4 method for DataMatrix
colAll(x, f, ..., na.rm = FALSE)

# S4 method for DataMatrix
rowAny(x, f, ..., na.rm = FALSE)

# S4 method for DataMatrix
colAny(x, f, ..., na.rm = FALSE)

Arguments

x, e1, e2
An object (typically a DataMatrix object).

na.rm
A logical scalar: should missing values (including NaN) be omitted from the calculations?

f
A logical predicate.

...
Further parameters to be passed to f.

Group Generics

DataMatrix objects have support for S4 group generic functionality to operate within elements across objects:

Arith

"+", "-", "*", "^", "%%", "%/%", "/"

Compare

"==", ">", "<", "!=", "<=", ">="

Logic

"&", "|"

Ops

"Arith", "Compare", "Logic"

Summary

"min", "max", "range", "prod", "sum", "any", "all"

Other Methods

rowAll and colAll

Check if all values are TRUE according to a given predicate.

rowAny and colAny

Check if some values are TRUE according to a given predicate.

rowCount and colCount

Count values in rows or columns according to a given predicate.

Examples

## Create count data matrix
A <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
B <- matrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Compare
A == B

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE FALSE FALSE FALSE FALSE
#>  [2,]  TRUE FALSE FALSE FALSE FALSE
#>  [3,]  TRUE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE FALSE  TRUE
#>  [6,] FALSE FALSE  TRUE FALSE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE
#>  [8,]  TRUE FALSE FALSE FALSE FALSE
#>  [9,] FALSE FALSE FALSE FALSE FALSE
#> [10,] FALSE FALSE FALSE FALSE FALSE
#> [11,] FALSE  TRUE FALSE FALSE FALSE
#> [12,] FALSE FALSE FALSE FALSE FALSE
#> [13,] FALSE FALSE FALSE FALSE FALSE
#> [14,]  TRUE FALSE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE FALSE FALSE
#> [16,] FALSE FALSE FALSE FALSE FALSE
#> [17,] FALSE FALSE FALSE FALSE FALSE
#> [18,] FALSE FALSE FALSE FALSE FALSE
#> [19,] FALSE FALSE FALSE FALSE FALSE
#> [20,] FALSE FALSE  TRUE FALSE FALSE

A != B

#>        [,1]  [,2]  [,3] [,4]  [,5]
#>  [1,]  TRUE  TRUE  TRUE TRUE  TRUE
#>  [2,] FALSE  TRUE  TRUE TRUE  TRUE
#>  [3,] FALSE  TRUE  TRUE TRUE  TRUE
#>  [4,]  TRUE  TRUE  TRUE TRUE  TRUE
#>  [5,]  TRUE  TRUE  TRUE TRUE FALSE
#>  [6,]  TRUE  TRUE FALSE TRUE  TRUE
#>  [7,]  TRUE  TRUE  TRUE TRUE  TRUE
#>  [8,] FALSE  TRUE  TRUE TRUE  TRUE
#>  [9,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [10,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [11,]  TRUE FALSE  TRUE TRUE  TRUE
#> [12,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [13,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [14,] FALSE  TRUE  TRUE TRUE  TRUE
#> [15,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [16,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [17,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [18,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [19,]  TRUE  TRUE  TRUE TRUE  TRUE
#> [20,]  TRUE  TRUE FALSE TRUE  TRUE

A > 5

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE FALSE  TRUE FALSE  TRUE
#>  [2,] FALSE  TRUE  TRUE FALSE FALSE
#>  [3,] FALSE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE  TRUE FALSE
#>  [5,] FALSE FALSE FALSE FALSE FALSE
#>  [6,]  TRUE  TRUE FALSE  TRUE FALSE
#>  [7,] FALSE  TRUE FALSE  TRUE  TRUE
#>  [8,] FALSE  TRUE FALSE FALSE FALSE
#>  [9,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [10,]  TRUE  TRUE  TRUE FALSE FALSE
#> [11,] FALSE  TRUE FALSE FALSE FALSE
#> [12,]  TRUE FALSE  TRUE FALSE  TRUE
#> [13,] FALSE FALSE FALSE FALSE  TRUE
#> [14,] FALSE  TRUE  TRUE FALSE  TRUE
#> [15,]  TRUE FALSE FALSE FALSE FALSE
#> [16,] FALSE  TRUE FALSE  TRUE FALSE
#> [17,] FALSE FALSE  TRUE  TRUE FALSE
#> [18,]  TRUE  TRUE  TRUE FALSE  TRUE
#> [19,]  TRUE FALSE FALSE FALSE FALSE
#> [20,]  TRUE FALSE  TRUE FALSE FALSE

A < 5

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,]  TRUE  TRUE FALSE  TRUE FALSE
#>  [2,]  TRUE FALSE FALSE  TRUE  TRUE
#>  [3,]  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [4,]  TRUE  TRUE  TRUE FALSE  TRUE
#>  [5,]  TRUE  TRUE  TRUE FALSE  TRUE
#>  [6,] FALSE FALSE  TRUE FALSE  TRUE
#>  [7,]  TRUE FALSE  TRUE FALSE FALSE
#>  [8,] FALSE FALSE  TRUE  TRUE  TRUE
#>  [9,] FALSE FALSE FALSE FALSE FALSE
#> [10,] FALSE FALSE FALSE  TRUE  TRUE
#> [11,]  TRUE FALSE  TRUE  TRUE  TRUE
#> [12,] FALSE  TRUE FALSE FALSE FALSE
#> [13,]  TRUE  TRUE  TRUE  TRUE FALSE
#> [14,]  TRUE FALSE FALSE FALSE FALSE
#> [15,] FALSE  TRUE  TRUE  TRUE  TRUE
#> [16,]  TRUE FALSE  TRUE FALSE  TRUE
#> [17,]  TRUE  TRUE FALSE FALSE  TRUE
#> [18,] FALSE FALSE FALSE  TRUE FALSE
#> [19,] FALSE  TRUE  TRUE  TRUE  TRUE
#> [20,] FALSE  TRUE FALSE  TRUE FALSE

A <= 10

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,] TRUE TRUE TRUE TRUE TRUE
#>  [2,] TRUE TRUE TRUE TRUE TRUE
#>  [3,] TRUE TRUE TRUE TRUE TRUE
#>  [4,] TRUE TRUE TRUE TRUE TRUE
#>  [5,] TRUE TRUE TRUE TRUE TRUE
#>  [6,] TRUE TRUE TRUE TRUE TRUE
#>  [7,] TRUE TRUE TRUE TRUE TRUE
#>  [8,] TRUE TRUE TRUE TRUE TRUE
#>  [9,] TRUE TRUE TRUE TRUE TRUE
#> [10,] TRUE TRUE TRUE TRUE TRUE
#> [11,] TRUE TRUE TRUE TRUE TRUE
#> [12,] TRUE TRUE TRUE TRUE TRUE
#> [13,] TRUE TRUE TRUE TRUE TRUE
#> [14,] TRUE TRUE TRUE TRUE TRUE
#> [15,] TRUE TRUE TRUE TRUE TRUE
#> [16,] TRUE TRUE TRUE TRUE TRUE
#> [17,] TRUE TRUE TRUE TRUE TRUE
#> [18,] TRUE TRUE TRUE TRUE TRUE
#> [19,] TRUE TRUE TRUE TRUE TRUE
#> [20,] TRUE TRUE TRUE TRUE TRUE

A >= 10

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,] FALSE FALSE FALSE FALSE FALSE
#>  [2,] FALSE FALSE FALSE FALSE FALSE
#>  [3,] FALSE FALSE FALSE FALSE FALSE
#>  [4,] FALSE FALSE FALSE FALSE FALSE
#>  [5,] FALSE FALSE FALSE FALSE FALSE
#>  [6,] FALSE FALSE FALSE FALSE FALSE
#>  [7,] FALSE FALSE FALSE FALSE FALSE
#>  [8,] FALSE FALSE FALSE FALSE FALSE
#>  [9,] FALSE  TRUE FALSE FALSE FALSE
#> [10,] FALSE FALSE FALSE FALSE FALSE
#> [11,] FALSE FALSE FALSE FALSE FALSE
#> [12,] FALSE FALSE FALSE FALSE FALSE
#> [13,] FALSE FALSE FALSE FALSE FALSE
#> [14,] FALSE FALSE FALSE FALSE FALSE
#> [15,] FALSE FALSE FALSE FALSE FALSE
#> [16,] FALSE  TRUE FALSE  TRUE FALSE
#> [17,] FALSE FALSE FALSE FALSE FALSE
#> [18,]  TRUE FALSE FALSE FALSE FALSE
#> [19,] FALSE FALSE FALSE FALSE FALSE
#> [20,] FALSE FALSE FALSE FALSE FALSE

## Arith
A + B

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    9    8   13    3   12
#>  [2,]    2    6    8    5    3
#>  [3,]    0    4    4   10    7
#>  [4,]    5    9   10    9    4
#>  [5,]    2    9    9    8    0
#>  [6,]   13   17    4   16    4
#>  [7,]    6   14   13    9   15
#>  [8,]   10   16    3   10    6
#>  [9,]   14   18   10    7   11
#> [10,]   11   10   15    5    8
#> [11,]    7   16    4    9    5
#> [12,]   18    5    9    9   10
#> [13,]    4   12   10    8   19
#> [14,]    0    6   11    8    8
#> [15,]   12    4    8    5    9
#> [16,]    7   18    8   15    5
#> [17,]    8    9   11    7    5
#> [18,]   10    9    9    5    9
#> [19,]   11    2    3    9   14
#> [20,]   15    7   12    8   13

A * 2

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    4    6   12    2   18
#>  [2,]    2   12   12    2    6
#>  [3,]    0    8    8    4    2
#>  [4,]    6    8    0   18    0
#>  [5,]    4    4    2   10    0
#>  [6,]   14   16    4   14    0
#>  [7,]    0   16    6   18   12
#>  [8,]   10   14    6    8    0
#>  [9,]   18   20   12   12   18
#> [10,]   16   14   18    2    0
#> [11,]    6   16    6    0    2
#> [12,]   16    0   12   10   14
#> [13,]    2    6    6    0   18
#> [14,]    0   12   16   10   16
#> [15,]   14    2    0    4    4
#> [16,]    8   20    4   20    4
#> [17,]    4    0   14   12    8
#> [18,]   20   18   18    4   12
#> [19,]   16    0    0    8    8
#> [20,]   18    6   12    0   10

A ^ 2

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    4    9   36    1   81
#>  [2,]    1   36   36    1    9
#>  [3,]    0   16   16    4    1
#>  [4,]    9   16    0   81    0
#>  [5,]    4    4    1   25    0
#>  [6,]   49   64    4   49    0
#>  [7,]    0   64    9   81   36
#>  [8,]   25   49    9   16    0
#>  [9,]   81  100   36   36   81
#> [10,]   64   49   81    1    0
#> [11,]    9   64    9    0    1
#> [12,]   64    0   36   25   49
#> [13,]    1    9    9    0   81
#> [14,]    0   36   64   25   64
#> [15,]   49    1    0    4    4
#> [16,]   16  100    4  100    4
#> [17,]    4    0   49   36   16
#> [18,]  100   81   81    4   36
#> [19,]   64    0    0   16   16
#> [20,]   81    9   36    0   25

## Summary
max(A)

#> [1] 10

min(A)

#> [1] 0

range(A)

#> [1]  0 10

prod(A)

#> [1] 0

sum(A)

#> [1] 427

any(A)

#> [1] TRUE

all(A)

#> [1] FALSE

## Logic
A & B

#>        [,1]  [,2]  [,3]  [,4]  [,5]
#>  [1,]  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [2,]  TRUE FALSE  TRUE  TRUE FALSE
#>  [3,] FALSE FALSE FALSE  TRUE  TRUE
#>  [4,]  TRUE  TRUE FALSE FALSE FALSE
#>  [5,] FALSE  TRUE  TRUE  TRUE FALSE
#>  [6,]  TRUE  TRUE  TRUE  TRUE FALSE
#>  [7,] FALSE  TRUE  TRUE FALSE  TRUE
#>  [8,]  TRUE  TRUE FALSE  TRUE FALSE
#>  [9,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [10,]  TRUE  TRUE  TRUE  TRUE FALSE
#> [11,]  TRUE  TRUE  TRUE FALSE  TRUE
#> [12,]  TRUE FALSE  TRUE  TRUE  TRUE
#> [13,]  TRUE  TRUE  TRUE FALSE  TRUE
#> [14,] FALSE FALSE  TRUE  TRUE FALSE
#> [15,]  TRUE  TRUE FALSE  TRUE  TRUE
#> [16,]  TRUE  TRUE  TRUE  TRUE  TRUE
#> [17,]  TRUE FALSE  TRUE  TRUE  TRUE
#> [18,] FALSE FALSE FALSE  TRUE  TRUE
#> [19,]  TRUE FALSE FALSE  TRUE  TRUE
#> [20,]  TRUE  TRUE  TRUE FALSE  TRUE

A | B

#>        [,1] [,2] [,3] [,4]  [,5]
#>  [1,]  TRUE TRUE TRUE TRUE  TRUE
#>  [2,]  TRUE TRUE TRUE TRUE  TRUE
#>  [3,] FALSE TRUE TRUE TRUE  TRUE
#>  [4,]  TRUE TRUE TRUE TRUE  TRUE
#>  [5,]  TRUE TRUE TRUE TRUE FALSE
#>  [6,]  TRUE TRUE TRUE TRUE  TRUE
#>  [7,]  TRUE TRUE TRUE TRUE  TRUE
#>  [8,]  TRUE TRUE TRUE TRUE  TRUE
#>  [9,]  TRUE TRUE TRUE TRUE  TRUE
#> [10,]  TRUE TRUE TRUE TRUE  TRUE
#> [11,]  TRUE TRUE TRUE TRUE  TRUE
#> [12,]  TRUE TRUE TRUE TRUE  TRUE
#> [13,]  TRUE TRUE TRUE TRUE  TRUE
#> [14,] FALSE TRUE TRUE TRUE  TRUE
#> [15,]  TRUE TRUE TRUE TRUE  TRUE
#> [16,]  TRUE TRUE TRUE TRUE  TRUE
#> [17,]  TRUE TRUE TRUE TRUE  TRUE
#> [18,]  TRUE TRUE TRUE TRUE  TRUE
#> [19,]  TRUE TRUE TRUE TRUE  TRUE
#> [20,]  TRUE TRUE TRUE TRUE  TRUE