Common Operations on Matrix Objects

Performs common operations on DataMatrix objects.

# S4 method for DataMatrix,DataMatrix
Arith(e1, e2)

# S4 method for DataMatrix,matrix
Arith(e1, e2)

# S4 method for DataMatrix,numeric
Arith(e1, e2)

# S4 method for DataMatrix,logical
Arith(e1, e2)

# S4 method for DataMatrix,DataMatrix
Compare(e1, e2)

# S4 method for DataMatrix,matrix
Compare(e1, e2)

# S4 method for DataMatrix,numeric
Compare(e1, e2)

# S4 method for DataMatrix,DataMatrix
Logic(e1, e2)

# S4 method for DataMatrix,matrix
Logic(e1, e2)

# S4 method for DataMatrix,numeric
Logic(e1, e2)

# S4 method for DataMatrix,logical
Logic(e1, e2)

# S4 method for DataMatrix
Math(x)

# S4 method for DataMatrix
Math2(x, digits)

# S4 method for DataMatrix
Summary(x, ..., na.rm = FALSE)

Arguments

x, e1, e2
An object (typically a DataMatrix object).

digits
A numeric value giving number of digits to be used in round or signif.

...
Further arguments passed to or from methods.

na.rm
A logical scalar: should missing values be removed?

Group Generics

DataMatrix objects have support for S4 group generic functionality to operate within elements across objects:

Arith

"+", "-", "*", "^", "%%", "%/%", "/"

Compare

"==", ">", "<", "!=", "<=", ">="

Logic

"&", "|"

Ops

"Arith", "Compare", "Logic"

Math

"abs", "sign", "sqrt", "ceiling", "floor", "trunc", "cummax", "cummin", "cumprod", "cumsum", "log", "log10", "log2", "log1p", "acos", "acosh", "asin", "asinh", "atan", "atanh", "exp", "expm1", "cos", "cosh", "cospi", "sin", "sinh", "sinpi", "tan", "tanh", "tanpi", "gamma", "lgamma", "digamma", "trigamma"

Math2

"round", "signif"

Summary

"min", "max", "range", "prod", "sum", "any", "all"

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Arith
A + B

#>       col1 col2 col3 col4 col5
#> row1     5   17   13    5   10
#> row2    13    9   16   11   14
#> row3    11   10    8    7    6
#> row4    12   12   11   12   15
#> row5     1    9   20   17    8
#> row6    16   17    1   13   11
#> row7     4   12   10    3   11
#> row8    14    3    6    9   18
#> row9     0    5    7    4    0
#> row10   13   11   15    4   11
#> row11   18    8   11   15   12
#> row12    1    2   13    3   16
#> row13   14   14   15    5    3
#> row14    1   14   19   12   17
#> row15   11    2   14    9    4
#> row16   11    9   16    8   18
#> row17    7   12    7   13    7
#> row18   10   11   12   19   17
#> row19   11    5   16    6   13
#> row20   12   12   17   13    4

A * 2

#>       col1 col2 col3 col4 col5
#> row1     8   16   14    2    6
#> row2    20    4   14    8   10
#> row3     2   14   16   14    6
#> row4     4    6   10   10   14
#> row5     2   18   20   16   12
#> row6    18   16    0   18    6
#> row7     8   14   16    0    2
#> row8    12    0   12   12   18
#> row9     0    4    4    6    0
#> row10   10   12   14    8    4
#> row11   18    6   12   10    6
#> row12    0    0   16    2   20
#> row13   12   18   18    8    4
#> row14    2    8   20    4   14
#> row15   12    0   20    6    4
#> row16    6    0   18    2   20
#> row17   12   18    4   18   14
#> row18    4   12    8   20   16
#> row19    4   10   12   12   10
#> row20   20   16   18   14    2

A ^ 2

#>       col1 col2 col3 col4 col5
#> row1    16   64   49    1    9
#> row2   100    4   49   16   25
#> row3     1   49   64   49    9
#> row4     4    9   25   25   49
#> row5     1   81  100   64   36
#> row6    81   64    0   81    9
#> row7    16   49   64    0    1
#> row8    36    0   36   36   81
#> row9     0    4    4    9    0
#> row10   25   36   49   16    4
#> row11   81    9   36   25    9
#> row12    0    0   64    1  100
#> row13   36   81   81   16    4
#> row14    1   16  100    4   49
#> row15   36    0  100    9    4
#> row16    9    0   81    1  100
#> row17   36   81    4   81   49
#> row18    4   36   16  100   64
#> row19    4   25   36   36   25
#> row20  100   64   81   49    1

## Compare
A == B

#>        col1  col2  col3  col4  col5
#> row1  FALSE FALSE FALSE FALSE FALSE
#> row2  FALSE FALSE FALSE FALSE FALSE
#> row3  FALSE FALSE FALSE FALSE  TRUE
#> row4  FALSE FALSE FALSE FALSE FALSE
#> row5  FALSE FALSE  TRUE FALSE FALSE
#> row6  FALSE FALSE FALSE FALSE FALSE
#> row7  FALSE FALSE FALSE FALSE FALSE
#> row8  FALSE FALSE FALSE FALSE  TRUE
#> row9   TRUE FALSE FALSE FALSE  TRUE
#> row10 FALSE FALSE FALSE FALSE FALSE
#> row11  TRUE FALSE FALSE FALSE FALSE
#> row12 FALSE FALSE FALSE FALSE FALSE
#> row13 FALSE FALSE FALSE FALSE FALSE
#> row14 FALSE FALSE FALSE FALSE FALSE
#> row15 FALSE FALSE FALSE FALSE  TRUE
#> row16 FALSE FALSE FALSE FALSE FALSE
#> row17 FALSE FALSE FALSE FALSE FALSE
#> row18 FALSE FALSE FALSE FALSE FALSE
#> row19 FALSE FALSE FALSE FALSE FALSE
#> row20 FALSE FALSE FALSE FALSE FALSE

A != B

#>        col1 col2  col3 col4  col5
#> row1   TRUE TRUE  TRUE TRUE  TRUE
#> row2   TRUE TRUE  TRUE TRUE  TRUE
#> row3   TRUE TRUE  TRUE TRUE FALSE
#> row4   TRUE TRUE  TRUE TRUE  TRUE
#> row5   TRUE TRUE FALSE TRUE  TRUE
#> row6   TRUE TRUE  TRUE TRUE  TRUE
#> row7   TRUE TRUE  TRUE TRUE  TRUE
#> row8   TRUE TRUE  TRUE TRUE FALSE
#> row9  FALSE TRUE  TRUE TRUE FALSE
#> row10  TRUE TRUE  TRUE TRUE  TRUE
#> row11 FALSE TRUE  TRUE TRUE  TRUE
#> row12  TRUE TRUE  TRUE TRUE  TRUE
#> row13  TRUE TRUE  TRUE TRUE  TRUE
#> row14  TRUE TRUE  TRUE TRUE  TRUE
#> row15  TRUE TRUE  TRUE TRUE FALSE
#> row16  TRUE TRUE  TRUE TRUE  TRUE
#> row17  TRUE TRUE  TRUE TRUE  TRUE
#> row18  TRUE TRUE  TRUE TRUE  TRUE
#> row19  TRUE TRUE  TRUE TRUE  TRUE
#> row20  TRUE TRUE  TRUE TRUE  TRUE

A > 5

#>        col1  col2  col3  col4  col5
#> row1  FALSE  TRUE  TRUE FALSE FALSE
#> row2   TRUE FALSE  TRUE FALSE FALSE
#> row3  FALSE  TRUE  TRUE  TRUE FALSE
#> row4  FALSE FALSE FALSE FALSE  TRUE
#> row5  FALSE  TRUE  TRUE  TRUE  TRUE
#> row6   TRUE  TRUE FALSE  TRUE FALSE
#> row7  FALSE  TRUE  TRUE FALSE FALSE
#> row8   TRUE FALSE  TRUE  TRUE  TRUE
#> row9  FALSE FALSE FALSE FALSE FALSE
#> row10 FALSE  TRUE  TRUE FALSE FALSE
#> row11  TRUE FALSE  TRUE FALSE FALSE
#> row12 FALSE FALSE  TRUE FALSE  TRUE
#> row13  TRUE  TRUE  TRUE FALSE FALSE
#> row14 FALSE FALSE  TRUE FALSE  TRUE
#> row15  TRUE FALSE  TRUE FALSE FALSE
#> row16 FALSE FALSE  TRUE FALSE  TRUE
#> row17  TRUE  TRUE FALSE  TRUE  TRUE
#> row18 FALSE  TRUE FALSE  TRUE  TRUE
#> row19 FALSE FALSE  TRUE  TRUE FALSE
#> row20  TRUE  TRUE  TRUE  TRUE FALSE

A < 5

#>        col1  col2  col3  col4  col5
#> row1   TRUE FALSE FALSE  TRUE  TRUE
#> row2  FALSE  TRUE FALSE  TRUE FALSE
#> row3   TRUE FALSE FALSE FALSE  TRUE
#> row4   TRUE  TRUE FALSE FALSE FALSE
#> row5   TRUE FALSE FALSE FALSE FALSE
#> row6  FALSE FALSE  TRUE FALSE  TRUE
#> row7   TRUE FALSE FALSE  TRUE  TRUE
#> row8  FALSE  TRUE FALSE FALSE FALSE
#> row9   TRUE  TRUE  TRUE  TRUE  TRUE
#> row10 FALSE FALSE FALSE  TRUE  TRUE
#> row11 FALSE  TRUE FALSE FALSE  TRUE
#> row12  TRUE  TRUE FALSE  TRUE FALSE
#> row13 FALSE FALSE FALSE  TRUE  TRUE
#> row14  TRUE  TRUE FALSE  TRUE FALSE
#> row15 FALSE  TRUE FALSE  TRUE  TRUE
#> row16  TRUE  TRUE FALSE  TRUE FALSE
#> row17 FALSE FALSE  TRUE FALSE FALSE
#> row18  TRUE FALSE  TRUE FALSE FALSE
#> row19  TRUE FALSE FALSE FALSE FALSE
#> row20 FALSE FALSE FALSE FALSE  TRUE

A <= 10

#>       col1 col2 col3 col4 col5
#> row1  TRUE TRUE TRUE TRUE TRUE
#> row2  TRUE TRUE TRUE TRUE TRUE
#> row3  TRUE TRUE TRUE TRUE TRUE
#> row4  TRUE TRUE TRUE TRUE TRUE
#> row5  TRUE TRUE TRUE TRUE TRUE
#> row6  TRUE TRUE TRUE TRUE TRUE
#> row7  TRUE TRUE TRUE TRUE TRUE
#> row8  TRUE TRUE TRUE TRUE TRUE
#> row9  TRUE TRUE TRUE TRUE TRUE
#> row10 TRUE TRUE TRUE TRUE TRUE
#> row11 TRUE TRUE TRUE TRUE TRUE
#> row12 TRUE TRUE TRUE TRUE TRUE
#> row13 TRUE TRUE TRUE TRUE TRUE
#> row14 TRUE TRUE TRUE TRUE TRUE
#> row15 TRUE TRUE TRUE TRUE TRUE
#> row16 TRUE TRUE TRUE TRUE TRUE
#> row17 TRUE TRUE TRUE TRUE TRUE
#> row18 TRUE TRUE TRUE TRUE TRUE
#> row19 TRUE TRUE TRUE TRUE TRUE
#> row20 TRUE TRUE TRUE TRUE TRUE

A >= 10

#>        col1  col2  col3  col4  col5
#> row1  FALSE FALSE FALSE FALSE FALSE
#> row2   TRUE FALSE FALSE FALSE FALSE
#> row3  FALSE FALSE FALSE FALSE FALSE
#> row4  FALSE FALSE FALSE FALSE FALSE
#> row5  FALSE FALSE  TRUE FALSE FALSE
#> row6  FALSE FALSE FALSE FALSE FALSE
#> row7  FALSE FALSE FALSE FALSE FALSE
#> row8  FALSE FALSE FALSE FALSE FALSE
#> row9  FALSE FALSE FALSE FALSE FALSE
#> row10 FALSE FALSE FALSE FALSE FALSE
#> row11 FALSE FALSE FALSE FALSE FALSE
#> row12 FALSE FALSE FALSE FALSE  TRUE
#> row13 FALSE FALSE FALSE FALSE FALSE
#> row14 FALSE FALSE  TRUE FALSE FALSE
#> row15 FALSE FALSE  TRUE FALSE FALSE
#> row16 FALSE FALSE FALSE FALSE  TRUE
#> row17 FALSE FALSE FALSE FALSE FALSE
#> row18 FALSE FALSE FALSE  TRUE FALSE
#> row19 FALSE FALSE FALSE FALSE FALSE
#> row20  TRUE FALSE FALSE FALSE FALSE

## Logic
A & B

#>        col1  col2  col3  col4  col5
#> row1   TRUE  TRUE  TRUE  TRUE  TRUE
#> row2   TRUE  TRUE  TRUE  TRUE  TRUE
#> row3   TRUE  TRUE FALSE FALSE  TRUE
#> row4   TRUE  TRUE  TRUE  TRUE  TRUE
#> row5  FALSE FALSE  TRUE  TRUE  TRUE
#> row6   TRUE  TRUE FALSE  TRUE  TRUE
#> row7  FALSE  TRUE  TRUE FALSE  TRUE
#> row8   TRUE FALSE FALSE  TRUE  TRUE
#> row9  FALSE  TRUE  TRUE  TRUE FALSE
#> row10  TRUE  TRUE  TRUE FALSE  TRUE
#> row11  TRUE  TRUE  TRUE  TRUE  TRUE
#> row12 FALSE FALSE  TRUE  TRUE  TRUE
#> row13  TRUE  TRUE  TRUE  TRUE  TRUE
#> row14 FALSE  TRUE  TRUE  TRUE  TRUE
#> row15  TRUE FALSE  TRUE  TRUE  TRUE
#> row16  TRUE FALSE  TRUE  TRUE  TRUE
#> row17  TRUE  TRUE  TRUE  TRUE FALSE
#> row18  TRUE  TRUE  TRUE  TRUE  TRUE
#> row19  TRUE FALSE  TRUE FALSE  TRUE
#> row20  TRUE  TRUE  TRUE  TRUE  TRUE

A | B

#>        col1 col2 col3 col4  col5
#> row1   TRUE TRUE TRUE TRUE  TRUE
#> row2   TRUE TRUE TRUE TRUE  TRUE
#> row3   TRUE TRUE TRUE TRUE  TRUE
#> row4   TRUE TRUE TRUE TRUE  TRUE
#> row5   TRUE TRUE TRUE TRUE  TRUE
#> row6   TRUE TRUE TRUE TRUE  TRUE
#> row7   TRUE TRUE TRUE TRUE  TRUE
#> row8   TRUE TRUE TRUE TRUE  TRUE
#> row9  FALSE TRUE TRUE TRUE FALSE
#> row10  TRUE TRUE TRUE TRUE  TRUE
#> row11  TRUE TRUE TRUE TRUE  TRUE
#> row12  TRUE TRUE TRUE TRUE  TRUE
#> row13  TRUE TRUE TRUE TRUE  TRUE
#> row14  TRUE TRUE TRUE TRUE  TRUE
#> row15  TRUE TRUE TRUE TRUE  TRUE
#> row16  TRUE TRUE TRUE TRUE  TRUE
#> row17  TRUE TRUE TRUE TRUE  TRUE
#> row18  TRUE TRUE TRUE TRUE  TRUE
#> row19  TRUE TRUE TRUE TRUE  TRUE
#> row20  TRUE TRUE TRUE TRUE  TRUE

## Math
sqrt(A)

#>           col1     col2     col3     col4     col5
#> row1  2.000000 2.828427 2.645751 1.000000 1.732051
#> row2  3.162278 1.414214 2.645751 2.000000 2.236068
#> row3  1.000000 2.645751 2.828427 2.645751 1.732051
#> row4  1.414214 1.732051 2.236068 2.236068 2.645751
#> row5  1.000000 3.000000 3.162278 2.828427 2.449490
#> row6  3.000000 2.828427 0.000000 3.000000 1.732051
#> row7  2.000000 2.645751 2.828427 0.000000 1.000000
#> row8  2.449490 0.000000 2.449490 2.449490 3.000000
#> row9  0.000000 1.414214 1.414214 1.732051 0.000000
#> row10 2.236068 2.449490 2.645751 2.000000 1.414214
#> row11 3.000000 1.732051 2.449490 2.236068 1.732051
#> row12 0.000000 0.000000 2.828427 1.000000 3.162278
#> row13 2.449490 3.000000 3.000000 2.000000 1.414214
#> row14 1.000000 2.000000 3.162278 1.414214 2.645751
#> row15 2.449490 0.000000 3.162278 1.732051 1.414214
#> row16 1.732051 0.000000 3.000000 1.000000 3.162278
#> row17 2.449490 3.000000 1.414214 3.000000 2.645751
#> row18 1.414214 2.449490 2.000000 3.162278 2.828427
#> row19 1.414214 2.236068 2.449490 2.449490 2.236068
#> row20 3.162278 2.828427 3.000000 2.645751 1.000000

## Summary
max(A)

#> [1] 10

min(A)

#> [1] 0

range(A)

#> [1]  0 10

prod(A)

#> [1] 0

sum(A)

#> [1] 505

any(A)

#> [1] TRUE

all(A)

#> [1] FALSE