Common Operations on Matrix Objects

Performs common operations on DataMatrix objects.

# S4 method for DataMatrix,DataMatrix
Arith(e1, e2)

# S4 method for DataMatrix,numeric
Arith(e1, e2)

# S4 method for DataMatrix,logical
Arith(e1, e2)

# S4 method for DataMatrix,DataMatrix
Compare(e1, e2)

# S4 method for DataMatrix,numeric
Compare(e1, e2)

# S4 method for DataMatrix,DataMatrix
Logic(e1, e2)

# S4 method for DataMatrix,numeric
Logic(e1, e2)

# S4 method for DataMatrix,logical
Logic(e1, e2)

# S4 method for DataMatrix
Math(x)

# S4 method for DataMatrix
Math2(x, digits)

# S4 method for DataMatrix
Summary(x, ..., na.rm = FALSE)

Arguments

x, e1, e2
An object (typically a DataMatrix object).

digits
A numeric value giving number of digits to be used in round or signif.

...
Further arguments passed to or from methods.

na.rm
A logical scalar: should missing values be removed?

Group Generics

DataMatrix objects have support for S4 group generic functionality to operate within elements across objects:

Arith

"+", "-", "*", "^", "%%", "%/%", "/"

Compare

"==", ">", "<", "!=", "<=", ">="

Logic

"&", "|"

Ops

"Arith", "Compare", "Logic"

Math

"abs", "sign", "sqrt", "ceiling", "floor", "trunc", "cummax", "cummin", "cumprod", "cumsum", "log", "log10", "log2", "log1p", "acos", "acosh", "asin", "asinh", "atan", "atanh", "exp", "expm1", "cos", "cosh", "cospi", "sin", "sinh", "sinpi", "tan", "tanh", "tanpi", "gamma", "lgamma", "digamma", "trigamma"

Math2

"round", "signif"

Summary

"min", "max", "range", "prod", "sum", "any", "all"

Author

N. Frerebeau

Examples

## Create count data matrix
A <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)
B <- CountMatrix(data = sample(0:10, 100, TRUE), nrow = 20, ncol = 5)

## Arith
A + B

#>       col1 col2 col3 col4 col5
#> row1    17    1   13   11   11
#> row2    12   10    3   11    2
#> row3     3    6    9   18    9
#> row4     5    7    4    0    0
#> row5    11   15    4   11   11
#> row6     8   11   15   12   11
#> row7     2   13    3   16    8
#> row8    14   15    5    3    8
#> row9    14   19   12   17    5
#> row10    2   14    9    4   14
#> row11    9   16    8   18   16
#> row12   12    7   13    7    4
#> row13   11   12   19   17   16
#> row14    5   16    6   13   10
#> row15   12   17   13    4    2
#> row16   13    5   10   11   16
#> row17   16   11   14    4   11
#> row18    8    7    6   16    5
#> row19   11   12   15   14   19
#> row20   20   17    8    2    8

A * 2

#>       col1 col2 col3 col4 col5
#> row1    16    0   18    6   14
#> row2    14   16    0    2    0
#> row3     0   12   12   18   16
#> row4     4    4    6    0    0
#> row5    12   14    8    4   16
#> row6     6   12   10    6   18
#> row7     0   16    2   20    2
#> row8    18   18    8    4   16
#> row9     8   20    4   14    0
#> row10    0   20    6    4   10
#> row11    0   18    2   20   16
#> row12   18    4   18   14    2
#> row13   12    8   20   16   16
#> row14   10   12   12   10   18
#> row15   16   18   14    2    4
#> row16   14    2    6    2   18
#> row17   14    8   10    6   14
#> row18   16   14    6   20    6
#> row19   10   10   14   20   18
#> row20   20   16   12    0    0

A ^ 2

#>       col1 col2 col3 col4 col5
#> row1    64    0   81    9   49
#> row2    49   64    0    1    0
#> row3     0   36   36   81   64
#> row4     4    4    9    0    0
#> row5    36   49   16    4   64
#> row6     9   36   25    9   81
#> row7     0   64    1  100    1
#> row8    81   81   16    4   64
#> row9    16  100    4   49    0
#> row10    0  100    9    4   25
#> row11    0   81    1  100   64
#> row12   81    4   81   49    1
#> row13   36   16  100   64   64
#> row14   25   36   36   25   81
#> row15   64   81   49    1    4
#> row16   49    1    9    1   81
#> row17   49   16   25    9   49
#> row18   64   49    9  100    9
#> row19   25   25   49  100   81
#> row20  100   64   36    0    0

## Compare
A == B

#>        col1  col2  col3  col4  col5
#> row1  FALSE FALSE FALSE FALSE FALSE
#> row2  FALSE FALSE FALSE FALSE FALSE
#> row3  FALSE FALSE FALSE  TRUE FALSE
#> row4  FALSE FALSE FALSE  TRUE  TRUE
#> row5  FALSE FALSE FALSE FALSE FALSE
#> row6  FALSE FALSE FALSE FALSE FALSE
#> row7  FALSE FALSE FALSE FALSE FALSE
#> row8  FALSE FALSE FALSE FALSE FALSE
#> row9  FALSE FALSE FALSE FALSE FALSE
#> row10 FALSE FALSE FALSE  TRUE FALSE
#> row11 FALSE FALSE FALSE FALSE  TRUE
#> row12 FALSE FALSE FALSE FALSE FALSE
#> row13 FALSE FALSE FALSE FALSE  TRUE
#> row14 FALSE FALSE FALSE FALSE FALSE
#> row15 FALSE FALSE FALSE FALSE FALSE
#> row16 FALSE FALSE FALSE FALSE FALSE
#> row17 FALSE FALSE FALSE FALSE FALSE
#> row18 FALSE FALSE  TRUE FALSE FALSE
#> row19 FALSE FALSE FALSE FALSE FALSE
#> row20  TRUE FALSE FALSE FALSE FALSE

A != B

#>        col1 col2  col3  col4  col5
#> row1   TRUE TRUE  TRUE  TRUE  TRUE
#> row2   TRUE TRUE  TRUE  TRUE  TRUE
#> row3   TRUE TRUE  TRUE FALSE  TRUE
#> row4   TRUE TRUE  TRUE FALSE FALSE
#> row5   TRUE TRUE  TRUE  TRUE  TRUE
#> row6   TRUE TRUE  TRUE  TRUE  TRUE
#> row7   TRUE TRUE  TRUE  TRUE  TRUE
#> row8   TRUE TRUE  TRUE  TRUE  TRUE
#> row9   TRUE TRUE  TRUE  TRUE  TRUE
#> row10  TRUE TRUE  TRUE FALSE  TRUE
#> row11  TRUE TRUE  TRUE  TRUE FALSE
#> row12  TRUE TRUE  TRUE  TRUE  TRUE
#> row13  TRUE TRUE  TRUE  TRUE FALSE
#> row14  TRUE TRUE  TRUE  TRUE  TRUE
#> row15  TRUE TRUE  TRUE  TRUE  TRUE
#> row16  TRUE TRUE  TRUE  TRUE  TRUE
#> row17  TRUE TRUE  TRUE  TRUE  TRUE
#> row18  TRUE TRUE FALSE  TRUE  TRUE
#> row19  TRUE TRUE  TRUE  TRUE  TRUE
#> row20 FALSE TRUE  TRUE  TRUE  TRUE

A > 5

#>        col1  col2  col3  col4  col5
#> row1   TRUE FALSE  TRUE FALSE  TRUE
#> row2   TRUE  TRUE FALSE FALSE FALSE
#> row3  FALSE  TRUE  TRUE  TRUE  TRUE
#> row4  FALSE FALSE FALSE FALSE FALSE
#> row5   TRUE  TRUE FALSE FALSE  TRUE
#> row6  FALSE  TRUE FALSE FALSE  TRUE
#> row7  FALSE  TRUE FALSE  TRUE FALSE
#> row8   TRUE  TRUE FALSE FALSE  TRUE
#> row9  FALSE  TRUE FALSE  TRUE FALSE
#> row10 FALSE  TRUE FALSE FALSE FALSE
#> row11 FALSE  TRUE FALSE  TRUE  TRUE
#> row12  TRUE FALSE  TRUE  TRUE FALSE
#> row13  TRUE FALSE  TRUE  TRUE  TRUE
#> row14 FALSE  TRUE  TRUE FALSE  TRUE
#> row15  TRUE  TRUE  TRUE FALSE FALSE
#> row16  TRUE FALSE FALSE FALSE  TRUE
#> row17  TRUE FALSE FALSE FALSE  TRUE
#> row18  TRUE  TRUE FALSE  TRUE FALSE
#> row19 FALSE FALSE  TRUE  TRUE  TRUE
#> row20  TRUE  TRUE  TRUE FALSE FALSE

A < 5

#>        col1  col2  col3  col4  col5
#> row1  FALSE  TRUE FALSE  TRUE FALSE
#> row2  FALSE FALSE  TRUE  TRUE  TRUE
#> row3   TRUE FALSE FALSE FALSE FALSE
#> row4   TRUE  TRUE  TRUE  TRUE  TRUE
#> row5  FALSE FALSE  TRUE  TRUE FALSE
#> row6   TRUE FALSE FALSE  TRUE FALSE
#> row7   TRUE FALSE  TRUE FALSE  TRUE
#> row8  FALSE FALSE  TRUE  TRUE FALSE
#> row9   TRUE FALSE  TRUE FALSE  TRUE
#> row10  TRUE FALSE  TRUE  TRUE FALSE
#> row11  TRUE FALSE  TRUE FALSE FALSE
#> row12 FALSE  TRUE FALSE FALSE  TRUE
#> row13 FALSE  TRUE FALSE FALSE FALSE
#> row14 FALSE FALSE FALSE FALSE FALSE
#> row15 FALSE FALSE FALSE  TRUE  TRUE
#> row16 FALSE  TRUE  TRUE  TRUE FALSE
#> row17 FALSE  TRUE FALSE  TRUE FALSE
#> row18 FALSE FALSE  TRUE FALSE  TRUE
#> row19 FALSE FALSE FALSE FALSE FALSE
#> row20 FALSE FALSE FALSE  TRUE  TRUE

A <= 10

#>       col1 col2 col3 col4 col5
#> row1  TRUE TRUE TRUE TRUE TRUE
#> row2  TRUE TRUE TRUE TRUE TRUE
#> row3  TRUE TRUE TRUE TRUE TRUE
#> row4  TRUE TRUE TRUE TRUE TRUE
#> row5  TRUE TRUE TRUE TRUE TRUE
#> row6  TRUE TRUE TRUE TRUE TRUE
#> row7  TRUE TRUE TRUE TRUE TRUE
#> row8  TRUE TRUE TRUE TRUE TRUE
#> row9  TRUE TRUE TRUE TRUE TRUE
#> row10 TRUE TRUE TRUE TRUE TRUE
#> row11 TRUE TRUE TRUE TRUE TRUE
#> row12 TRUE TRUE TRUE TRUE TRUE
#> row13 TRUE TRUE TRUE TRUE TRUE
#> row14 TRUE TRUE TRUE TRUE TRUE
#> row15 TRUE TRUE TRUE TRUE TRUE
#> row16 TRUE TRUE TRUE TRUE TRUE
#> row17 TRUE TRUE TRUE TRUE TRUE
#> row18 TRUE TRUE TRUE TRUE TRUE
#> row19 TRUE TRUE TRUE TRUE TRUE
#> row20 TRUE TRUE TRUE TRUE TRUE

A >= 10

#>        col1  col2  col3  col4  col5
#> row1  FALSE FALSE FALSE FALSE FALSE
#> row2  FALSE FALSE FALSE FALSE FALSE
#> row3  FALSE FALSE FALSE FALSE FALSE
#> row4  FALSE FALSE FALSE FALSE FALSE
#> row5  FALSE FALSE FALSE FALSE FALSE
#> row6  FALSE FALSE FALSE FALSE FALSE
#> row7  FALSE FALSE FALSE  TRUE FALSE
#> row8  FALSE FALSE FALSE FALSE FALSE
#> row9  FALSE  TRUE FALSE FALSE FALSE
#> row10 FALSE  TRUE FALSE FALSE FALSE
#> row11 FALSE FALSE FALSE  TRUE FALSE
#> row12 FALSE FALSE FALSE FALSE FALSE
#> row13 FALSE FALSE  TRUE FALSE FALSE
#> row14 FALSE FALSE FALSE FALSE FALSE
#> row15 FALSE FALSE FALSE FALSE FALSE
#> row16 FALSE FALSE FALSE FALSE FALSE
#> row17 FALSE FALSE FALSE FALSE FALSE
#> row18 FALSE FALSE FALSE  TRUE FALSE
#> row19 FALSE FALSE FALSE  TRUE FALSE
#> row20  TRUE FALSE FALSE FALSE FALSE

## Logic
A & B

#>        col1  col2  col3  col4  col5
#> row1   TRUE FALSE  TRUE  TRUE  TRUE
#> row2   TRUE  TRUE FALSE  TRUE FALSE
#> row3  FALSE FALSE  TRUE  TRUE  TRUE
#> row4   TRUE  TRUE  TRUE FALSE FALSE
#> row5   TRUE  TRUE FALSE  TRUE  TRUE
#> row6   TRUE  TRUE  TRUE  TRUE  TRUE
#> row7  FALSE  TRUE  TRUE  TRUE  TRUE
#> row8   TRUE  TRUE  TRUE  TRUE FALSE
#> row9   TRUE  TRUE  TRUE  TRUE FALSE
#> row10 FALSE  TRUE  TRUE  TRUE  TRUE
#> row11 FALSE  TRUE  TRUE  TRUE  TRUE
#> row12  TRUE  TRUE  TRUE FALSE  TRUE
#> row13  TRUE  TRUE  TRUE  TRUE  TRUE
#> row14 FALSE  TRUE FALSE  TRUE  TRUE
#> row15  TRUE  TRUE  TRUE  TRUE FALSE
#> row16  TRUE  TRUE  TRUE  TRUE  TRUE
#> row17  TRUE  TRUE  TRUE  TRUE  TRUE
#> row18 FALSE FALSE  TRUE  TRUE  TRUE
#> row19  TRUE  TRUE  TRUE  TRUE  TRUE
#> row20  TRUE  TRUE  TRUE FALSE FALSE

A | B

#>       col1 col2 col3  col4  col5
#> row1  TRUE TRUE TRUE  TRUE  TRUE
#> row2  TRUE TRUE TRUE  TRUE  TRUE
#> row3  TRUE TRUE TRUE  TRUE  TRUE
#> row4  TRUE TRUE TRUE FALSE FALSE
#> row5  TRUE TRUE TRUE  TRUE  TRUE
#> row6  TRUE TRUE TRUE  TRUE  TRUE
#> row7  TRUE TRUE TRUE  TRUE  TRUE
#> row8  TRUE TRUE TRUE  TRUE  TRUE
#> row9  TRUE TRUE TRUE  TRUE  TRUE
#> row10 TRUE TRUE TRUE  TRUE  TRUE
#> row11 TRUE TRUE TRUE  TRUE  TRUE
#> row12 TRUE TRUE TRUE  TRUE  TRUE
#> row13 TRUE TRUE TRUE  TRUE  TRUE
#> row14 TRUE TRUE TRUE  TRUE  TRUE
#> row15 TRUE TRUE TRUE  TRUE  TRUE
#> row16 TRUE TRUE TRUE  TRUE  TRUE
#> row17 TRUE TRUE TRUE  TRUE  TRUE
#> row18 TRUE TRUE TRUE  TRUE  TRUE
#> row19 TRUE TRUE TRUE  TRUE  TRUE
#> row20 TRUE TRUE TRUE  TRUE  TRUE

## Math
sqrt(A)

#>           col1     col2     col3     col4     col5
#> row1  2.828427 0.000000 3.000000 1.732051 2.645751
#> row2  2.645751 2.828427 0.000000 1.000000 0.000000
#> row3  0.000000 2.449490 2.449490 3.000000 2.828427
#> row4  1.414214 1.414214 1.732051 0.000000 0.000000
#> row5  2.449490 2.645751 2.000000 1.414214 2.828427
#> row6  1.732051 2.449490 2.236068 1.732051 3.000000
#> row7  0.000000 2.828427 1.000000 3.162278 1.000000
#> row8  3.000000 3.000000 2.000000 1.414214 2.828427
#> row9  2.000000 3.162278 1.414214 2.645751 0.000000
#> row10 0.000000 3.162278 1.732051 1.414214 2.236068
#> row11 0.000000 3.000000 1.000000 3.162278 2.828427
#> row12 3.000000 1.414214 3.000000 2.645751 1.000000
#> row13 2.449490 2.000000 3.162278 2.828427 2.828427
#> row14 2.236068 2.449490 2.449490 2.236068 3.000000
#> row15 2.828427 3.000000 2.645751 1.000000 1.414214
#> row16 2.645751 1.000000 1.732051 1.000000 3.000000
#> row17 2.645751 2.000000 2.236068 1.732051 2.645751
#> row18 2.828427 2.645751 1.732051 3.162278 1.732051
#> row19 2.236068 2.236068 2.645751 3.162278 3.000000
#> row20 3.162278 2.828427 2.449490 0.000000 0.000000

## Summary
max(A)

#> [1] 10

min(A)

#> [1] 0

range(A)

#> [1]  0 10

prod(A)

#> [1] 0

sum(A)

#> [1] 515

any(A)

#> [1] TRUE

all(A)

#> [1] FALSE